% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_exdata.R
\docType{data}
\name{berkeley_exdata}
\alias{berkeley_exdata}
\title{Berkeley Child Guidance Study Data for Females}
\format{
A data frame with the following 3 variables:
\describe{
\item{id}{A factor variable identifying the subject.}
\item{age}{Age in years, numeric vector.}
\item{height}{Height in centimeters, numeric vector.}
}
}
\usage{
berkeley_exdata
}
\value{
A data frame with 3 columns.
}
\description{
A subset of the \link{berkeley} data, containing
longitudinal growth data for 70 females (aged 8 to 18 years).
}
\details{
Detailed information about the full dataset can be found in the
\link{berkeley} data.
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
