% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_exfit.R
\docType{data}
\name{berkeley_exfit}
\alias{berkeley_exfit}
\title{Model Fit to the Berkeley Child Guidance Study Data for Females}
\format{
A model fit object containing a summary of the posterior draws.
}
\usage{
berkeley_exfit
}
\value{
An object of class \code{bgmfit} containing the posterior draws.
}
\description{
Bayesian SITAR model fit to the \link{berkeley_exdata}
dataset (70 females, aged 8 to 18 years).
}
\details{
Detailed information about the data can be found in the
\link{berkeley_exdata} dataset.
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
