% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_ffit.R
\docType{data}
\name{berkeley_ffit}
\alias{berkeley_ffit}
\title{Model fit to the Berkeley data for females}
\format{
Model fit comprising summary of posterior draws.
}
\usage{
berkeley_ffit
}
\value{
An object of class \code{bgmfit} with posterior draws.
}
\description{
Bayesian SITAR model fit to the \link{berkeley_fdata}
data (20 females, 8 to 18 years of age).
}
\details{
Data details are provided in the \link{berkeley_fdata}.
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
