% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_fit.R
\docType{data}
\name{berkeley_fit}
\alias{berkeley_fit}
\title{Model fit to the Berkeley data set}
\format{
A \code{bgmfit} object for the model fit to the Berkeley data.
}
\usage{
berkeley_fit
}
\value{
An object of class \code{bgmfit} with posterior draws.
}
\description{
Model is fit to the Berkeley Child Guidance Study dataset
containing longitudinal growth measurements made on 136 children from birth
to 21 years.
}
\details{
Data details are provided in the \link{berkeley} and also in
the \emph{sitar} package \insertCite{R-sitar}{bsitar}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
