% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_mdata.R
\docType{data}
\name{berkeley_mdata}
\alias{berkeley_mdata}
\title{Berkeley Child Guidance Study Data for males}
\format{
A data frame with 902 observations on the following 3 variables:
\describe{
\item{id}{factor with levels 201-278}
\item{age}{years, numeric vector}
\item{height}{cm, numeric vector}
}
}
\usage{
berkeley_mdata
}
\value{
A data frame with 3 columns.
}
\description{
A subset of the Berkeley Child Guidance Study data
\strong{berkeley} that contains longitudinal growth data for 66 males (6 to
20 years of age).
}
\details{
A detailed description of the full data including the frequency of
measurements per year is provided in the \link{berkeley} dataset.
}
\references{
\insertAllCited{}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
