% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_mfit.R
\docType{data}
\name{berkeley_mfit}
\alias{berkeley_mfit}
\title{Model fit to the Berkeley data for males}
\format{
An object of class \code{bgmfit} fit to the Berkeley data.
}
\usage{
berkeley_mfit
}
\value{
An object of class \code{bgmfit} with posterior draws.
}
\description{
Model is fit to the Berkeley Child Guidance Study dataset
containing longitudinal growth measurements made on 66 males between 6
and 20 years.
}
\details{
Data details are provided in the \link{berkeley_mdata}.
}
\references{
\insertAllCited{}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
