% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{compare_model.bgmfit}
\alias{compare_model.bgmfit}
\alias{compare_model}
\alias{compare_models}
\title{Model comparison with the loo package}
\usage{
\method{compare_model}{bgmfit}(
  model,
  ...,
  criterion = "loo",
  model_names = NULL,
  check_criterion = TRUE,
  add_criterion_args = list(),
  verbose = FALSE
)

compare_model(model, ...)

compare_models(model, ...)
}
\arguments{
\item{model}{An \code{R} object of class \code{bsitar}}

\item{...}{More \code{brmsfit} objects.}

\item{criterion}{The name of the criterion to be extracted
from \code{brmsfit} objects.}

\item{model_names}{If \code{NULL} (the default) will use model names
derived from deparsing the call. Otherwise will use the passed
values as model names.}

\item{check_criterion}{A logical (default \code{FALSE}) indicating whether to
check and add fit criterion if not already added to model objects}

\item{add_criterion_args}{A named list to set up the arguments for the
\code{\link[=add_model_criterion]{add_model_criterion()}} function. Ignored when \code{check_criterion =
  FALSE} or when each model objects has precomputed fit criteria.}

\item{verbose}{A logical (default \code{FALSE}) to print some useful
information.}
}
\value{
An object of class "\code{compare.loo}".
}
\description{
A wrapper around \code{\link[brms:loo_compare.brmsfit]{brms::loo_compare()}} to compare models using fit criteria.
}
\details{
All models should have pre computed fit criterion See \code{\link[=add_model_criterion]{add_model_criterion()}}
for more help.
}
\examples{
\donttest{
# Fit Bayesian SITAR model 

# To avoid model estimation which can take time, the Bayesian SITAR model fit
# to the 'berkeley_exdata' has been saved as an example fit ('berkeley_exfit').
# See 'bsitar' function for details on 'berkeley_exdata' and 'berkeley_exfit'.

model <- getNsObject(berkeley_exfit)

# Add model fit criteria (e.g., WAIC). For illustration purposes, we compare 
# model with itself
model_1 <- add_model_criterion(model, criterion = c("waic"))
model_2 <- add_model_criterion(model, criterion = c("waic"))

# compare models model_1 and model_2
compare_model_12 <- compare_model(model_1, model_2, criterion = c("waic"))
}

}
\seealso{
\code{\link[brms:loo_compare.brmsfit]{brms::loo_compare()}}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
