% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons_bgm.R
\name{comparisons_bgm.bgmfit}
\alias{comparisons_bgm.bgmfit}
\alias{comparisons_bgm}
\title{Compute growth parameters for the \code{bgmfit} model}
\usage{
\method{comparisons_bgm}{bgmfit}(
  model,
  resp = NULL,
  ndraws = NULL,
  draw_ids = NULL,
  newdata = NULL,
  re_formula = NA,
  parameter = NULL,
  xrange = 1,
  acg_velocity = 0.1,
  digits = 2,
  numeric_cov_at = NULL,
  aux_variables = NULL,
  levels_id = NULL,
  avg_reffects = NULL,
  ipts = NULL,
  seed = 123,
  future = FALSE,
  future_session = "multisession",
  cores = NULL,
  variables = NULL,
  comparison = "difference",
  type = NULL,
  vcov = TRUE,
  by = FALSE,
  conf_level = 0.95,
  transform = NULL,
  cross = FALSE,
  wts = NULL,
  hypothesis = NULL,
  equivalence = NULL,
  p_adjust = NULL,
  df = Inf,
  eps = NULL,
  ...
)

comparisons_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{resp}{An optional character string to specify response variable when
estimating growth parameter for the univariate-by-subgroup and multivariate
models (see \code{\link[=bgm]{bgm()}} for details on univariate-by-subgroup and
multivariate models). For univariate model, \code{resp = NULL} (default).}

\item{ndraws}{Positive integer indicating the number of posterior draws to be
used in estimation. If \code{NULL} (default), all draws are used.}

\item{draw_ids}{An integer indicating the specif posterior draw(s)
to be used. If \code{NULL} (default), all draws are used.}

\item{newdata}{An optional data frame to be used in predictions. If
\code{NULL} (default), the original data from the fitted model is used.}

\item{re_formula}{Option to indicate whether or not to include the
individual/group-level effects in the estimation. When \code{NA} (default),
the individual-level effects are excluded and therefore population average
growth parameters are computed. When \code{NULL}, individual-level effects
are included in the computation and hence the growth parameters estimates
returned are individual-specific. In both situations, (i.e,, \code{NA} or
\code{NULL}), continuous and factor covariate(s) are appropriately included
in the estimation. The continuous covariates by default are set to their
means (see \code{numeric_cov_at} for details) whereas factor covariates are
left unaltered thereby allowing estimation of (factor) covariate specific
population average and individual-specific growth parameter at mean value
of continous covaristes(s).}

\item{parameter}{A single character string, or a vector of character strings
indicating the parameter to be estimated. Options available are \code{ato},
\code{vto}, \code{vpv}, \code{apv}, \code{vcg}, and \code{acg}. For
\code{parameter = NULL} (default), age at peak growth velocity is estimated.}

\item{xrange}{An integer to set the predictor range (i.e., age) when
executing the interpolation via \code{ipts}. The default \code{NULL} sets
the individual specific predictor range whereas code \code{xrange = 1} sets
same range for all individuals within the higher order grouping variable
(e.g., study). Code \code{xrange  = 2} sets the identical range dplyr::across the
entire sample. Lastly, a paired numeric values can be supplied e.g.,
\code{xrange = c(6, 20)} will set the range between 6 and 20.}

\item{acg_velocity}{A real number to set the percentage of the peak growth
as the cessation velocity. The \code{acg_velocity} should be greater than
\code{0} and less than \code{1}. The default \code{acg_velocity = 0.10}
indicates that a 10 per cent of the peak growth velocity will be used to
get the cessation velocity and the age at the cessation velocity. For
example is peak growth velocity is estimated as \code{10 mm/year}, then
cessation velocity would be \code{1 mm/year}.}

\item{digits}{An integer (default \code{2}) to set the rounding disgits.}

\item{numeric_cov_at}{An optional argument to specify the value of continuous
covariate(s). The default \code{NULL} option set the continuous
covariate(s) at their mean. Alternatively, a named list can be
supplied to manually set these values. For example, \code{numeric_cov_at =
  list(xx = 2)} will set the continuous covariate varibale 'xx' at 2. The
argument \code{numeric_cov_at} is ignored when no continuous covariate is
included in the model.}

\item{aux_variables}{An optional argument to specify the variables that can be
passed to the \code{ipts} argument (see below). This is useful when fitting
location scale models and the measurement error models.}

\item{levels_id}{An optional argument to specify the ids for hierarchical
model (default \code{NULL}. It is used only when model is fitted to the
data with 3 or more levels of hierarchy. For a two level model, the id for
second level is automatically inferred from the fitted model. Even for 3 or
higher level model, ids are inferred from the fitted model but under the
assumption that hierarchy is specified from lower to upper levels i.e, id,
study assuming that id is nested within the studies. However, it is not
gauranted that these ids are sorted correctly. Therefore, it is better to
set them manually.}

\item{avg_reffects}{An optional argument (default \code{NULL} to calculate
(marginal/average) curves and growth parameters (such as APGV and PGV). If
specified, it must be a named list indicating the \code{over} and the fixed
efects \code{feby} and random effects efects \code{reby} arguments e.g.,
\code{avg_reffects = list(feby = 'study', reby = NULL, over = 'age'}. The
\code{over} is typically age and is used to average over the random
effects. The second argument is \code{by} that specifies the factor
variable by which \code{over} is executed.}

\item{ipts}{An integer to set the length of the predictor variable to get a
smooth velocity curve. The \code{NULL} (default) will return original
values whereas an integer (e.g., \code{ipts = 10}) will interpolate the
predictor. It is important to note that these interpolations do not alter
the range of predictor when calculating population average and the
individual specific velocity curves.}

\item{seed}{An integer (default \code{123}) that is passed to the estimation
method.}

\item{future}{A logical (default \code{FALSE}) to specify whether or not to
perform parallel computations. If set to \code{TRUE}, the
\code{\link[future.apply:future_lapply]{future.apply::future_sapply()}} function is used for summarizing the draws.}

\item{future_session}{A character string to set the session type when
\code{future = TRUE}. The \code{multisession} (default) options sets the
multisession whereas the \code{multicore} sets the multicore session. Note
that multicore session is not supported on Windows systems. For more
details, see \code{\link[future.apply:future_lapply]{future.apply::future_sapply()}}.}

\item{cores}{Number of cores to be used when running the parallel
computations by setting the option \code{future = TRUE}. On non-Windows,
systems this argument can be set globally via the mc.cores option. For the
default \code{NULL} option, the number of cores are set automatically by
calling the \code{\link[future:re-exports]{future::availableCores()}}. The number of cores used are the
maximum number of cores avaialble minus one, i.e.,
\code{future::availableCores() - 1}.}

\item{variables}{Focal variables
\itemize{
\item \code{NULL}: compute comparisons for all the variables in the model object (can be slow).
\item Character vector: subset of variables (usually faster).
\item Named list: names identify the subset of variables of interest, and values define the type of contrast to compute. Acceptable values depend on the variable type:
\itemize{
\item Factor or character variables:
\itemize{
\item "reference": Each factor level is compared to the factor reference (base) level
\item "all": All combinations of observed levels
\item "sequential": Each factor level is compared to the previous factor level
\item "pairwise": Each factor level is compared to all other levels
\item "minmax": The highest and lowest levels of a factor.
\item "revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses.
\item Vector of length 2 with the two values to compare.
}
\item Logical variables:
\itemize{
\item NULL: contrast between TRUE and FALSE
}
\item Numeric variables:
\itemize{
\item Numeric of length 1: Contrast for a gap of \code{x}, computed at the observed value plus and minus \code{x / 2}. For example, estimating a \code{+1} contrast compares adjusted predictions when the regressor is equal to its observed value minus 0.5 and its observed value plus 0.5.
\item Numeric of length equal to the number of rows in \code{newdata}: Same as above, but the contrast can be customized for each row of \code{newdata}.
\item Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the \code{x} vector.
\item Data frame with the same number of rows as \code{newdata}, with two columns of "low" and "high" values to compare.
\item Function which accepts a numeric vector and returns a data frame with two columns of "low" and "high" values to compare. See examples below.
\item "iqr": Contrast across the interquartile range of the regressor.
\item "sd": Contrast across one standard deviation around the regressor mean.
\item "2sd": Contrast across two standard deviations around the regressor mean.
\item "minmax": Contrast between the maximum and the minimum values of the regressor.
}
\item Examples:
\itemize{
\item \code{variables = list(gear = "pairwise", hp = 10)}
\item \code{variables = list(gear = "sequential", hp = c(100, 120))}
\item See the Examples section below for more.
}
}
}}

\item{comparison}{How should pairs of predictions be compared? Difference, ratio, odds ratio, or user-defined functions.
\itemize{
\item string: shortcuts to common contrast functions.
\itemize{
\item Supported shortcuts strings: difference, differenceavg, differenceavgwts, dydx, eyex, eydx, dyex, dydxavg, eyexavg, eydxavg, dyexavg, dydxavgwts, eyexavgwts, eydxavgwts, dyexavgwts, ratio, ratioavg, ratioavgwts, lnratio, lnratioavg, lnratioavgwts, lnor, lnoravg, lnoravgwts, lift, liftavg, expdydx, expdydxavg, expdydxavgwts
\item See the Comparisons section below for definitions of each transformation.
}
\item function: accept two equal-length numeric vectors of adjusted predictions (\code{hi} and \code{lo}) and returns a vector of contrasts of the same length, or a unique numeric value.
\itemize{
\item See the Transformations section below for examples of valid functions.
}
}}

\item{type}{string indicates the type (scale) of the predictions used to
compute contrasts or slopes. This can differ based on the model
type, but will typically be a string such as: "response", "link", "probs",
or "zero". When an unsupported string is entered, the model-specific list of
acceptable values is returned in an error message. When \code{type} is \code{NULL}, the
first entry in the error message is used by default.}

\item{vcov}{Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:
\itemize{
\item FALSE: Do not compute standard errors. This can speed up computation considerably.
\item TRUE: Unit-level standard errors using the default \code{vcov(model)} variance-covariance matrix.
\item String which indicates the kind of uncertainty estimates to return.
\itemize{
\item Heteroskedasticity-consistent: \code{"HC"}, \code{"HC0"}, \code{"HC1"}, \code{"HC2"}, \code{"HC3"}, \code{"HC4"}, \code{"HC4m"}, \code{"HC5"}. See \code{?sandwich::vcovHC}
\item Heteroskedasticity and autocorrelation consistent: \code{"HAC"}
\item Mixed-Models degrees of freedom: "satterthwaite", "kenward-roger"
\item Other: \code{"NeweyWest"}, \code{"KernHAC"}, \code{"OPG"}. See the \code{sandwich} package documentation.
}
\item One-sided formula which indicates the name of cluster variables (e.g., \code{~unit_id}). This formula is passed to the \code{cluster} argument of the \code{sandwich::vcovCL} function.
\item Square covariance matrix
\item Function which returns a covariance matrix (e.g., \code{stats::vcov(model)})
}}

\item{by}{Aggregate unit-level estimates (aka, marginalize, average over). Valid inputs:
\itemize{
\item \code{FALSE}: return the original unit-level estimates.
\item \code{TRUE}: aggregate estimates for each term.
\item Character vector of column names in \code{newdata} or in the data frame produced by calling the function without the \code{by} argument.
\item Data frame with a \code{by} column of group labels, and merging columns shared by \code{newdata} or the data frame produced by calling the same function without the \code{by} argument.
\item See examples below.
\item For more complex aggregations, you can use the \code{FUN} argument of the \code{hypotheses()} function. See that function's documentation and the Hypothesis Test vignettes on the \code{marginaleffects} website.
}}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval.}

\item{transform}{string or function. Transformation applied to unit-level estimates and confidence intervals just before the function returns results. Functions must accept a vector and return a vector of the same length. Support string shortcuts: "exp", "ln"}

\item{cross}{\itemize{
\item \code{FALSE}: Contrasts represent the change in adjusted predictions when one predictor changes and all other variables are held constant.
\item \code{TRUE}: Contrasts represent the changes in adjusted predictions when all the predictors specified in the \code{variables} argument are manipulated simultaneously (a "cross-contrast").
}}

\item{wts}{string or numeric: weights to use when computing average contrasts or slopes. These weights only affect the averaging in \verb{avg_*()} or with the \code{by} argument, and not the unit-level estimates themselves. Internally, estimates and weights are passed to the \code{weighted.mean()} function.
\itemize{
\item string: column name of the weights variable in \code{newdata}. When supplying a column name to \code{wts}, it is recommended to supply the original data (including the weights variable) explicitly to \code{newdata}.
\item numeric: vector of length equal to the number of rows in the original data or in \code{newdata} (if supplied).
}}

\item{hypothesis}{specify a hypothesis test or custom contrast using a numeric value, vector, or matrix, a string, or a string formula.
\itemize{
\item Numeric:
\itemize{
\item Single value: the null hypothesis used in the computation of Z and p (before applying \code{transform}).
\item Vector: Weights to compute a linear combination of (custom contrast between) estimates. Length equal to the number of rows generated by the same function call, but without the \code{hypothesis} argument.
\item Matrix: Each column is a vector of weights, as describe above, used to compute a distinct linear combination of (contrast between) estimates. The column names of the matrix are used as labels in the output.
}
\item String formula to specify linear or non-linear hypothesis tests. If the \code{term} column uniquely identifies rows, terms can be used in the formula. Otherwise, use \code{b1}, \code{b2}, etc. to identify the position of each parameter. The \verb{b*} wildcard can be used to test hypotheses on all estimates. Examples:
\itemize{
\item \code{hp = drat}
\item \code{hp + drat = 12}
\item \code{b1 + b2 + b3 = 0}
\item \verb{b* / b1 = 1}
}
\item String:
\itemize{
\item "pairwise": pairwise differences between estimates in each row.
\item "reference": differences between the estimates in each row and the estimate in the first row.
\item "sequential": difference between an estimate and the estimate in the next row.
\item "revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses, as described above.
}
\item See the Examples section below and the vignette: https://marginaleffects.com/articles/hypothesis.html
}}

\item{equivalence}{Numeric vector of length 2: bounds used for the two-one-sided test (TOST) of equivalence, and for the non-inferiority and non-superiority tests. See Details section below.}

\item{p_adjust}{Adjust p-values for multiple comparisons: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", or "fdr". See \link[stats:p.adjust]{stats::p.adjust}}

\item{df}{Degrees of freedom used to compute p values and confidence intervals. A single numeric value between 1 and \code{Inf}. When \code{df} is \code{Inf}, the normal distribution is used. When \code{df} is finite, the \code{t} distribution is used. See \link[insight:get_df]{insight::get_df} for a convenient function to extract degrees of freedom. Ex: \code{slopes(model, df = insight::get_df(model))}}

\item{eps}{NULL or numeric value which determines the step size to use when
calculating numerical derivatives: (f(x+eps)-f(x))/eps. When \code{eps} is
\code{NULL}, the step size is 0.0001 multiplied by the difference between
the maximum and minimum values of the variable with respect to which we
are taking the derivative. Changing \code{eps} may be necessary to avoid
numerical problems in certain models.}

\item{...}{Further arguments passed to \code{brms::fitted()} and
\code{brms::predict()} functions that control several aspects of data
validation and prediction. See \code{\link[brms:fitted.brmsfit]{brms::fitted.brmsfit()}} and
\code{\link[brms:predict.brmsfit]{brms::predict.brmsfit()}} for details.}
}
\value{
A data frame objects with estimates and CIs for computed parameter(s)
}
\description{
The \code{comparisons_bgm} provides estimates of growth
parameters (such as peak velocity and the age at peak velocity) derived
from the growth velocity curve.
}
\details{
The \code{comparisons_bgm} function estimates and returns the
following growth paramaeter:
\itemize{
\item pv  - peak velocity
\item apv - age at peak velocity
\item tv  - take off velocity
\item atv - age at take off velocity
\item cv  - cessation velocity
\item acv - age at cessation velocity
}

The take off velocity is the minimum velocity before the peak velocity and it
indicates the beginning of the pubertal growth spurt. The cessation velocity
indicates the end of the active pubertal growth spurt and is calculated as
the percentage of the peak velocity (\code{pv}). The percentage of the peak
velocity used in the calculation of the the cessation velocity (\code{'cv'})
is controlled via the \code{acg_velocity} argument. The \code{acg_velocity}
takes a real value as an input (greater than 0 and less than 1). Typically, a
10 percent of \code{pv} (i.e., \code{acg_velocity = 0.1}) is considered as a
good indicator of the cessation of the active pubertal growth spurt
(\insertCite{Anna2022}{bsitar}).
}
\examples{
\dontrun{
comparisons_bgm(model, parameter = 'apv')
}

}
\references{
\insertAllCited{}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
