% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects_.bsitar.R
\name{conditional_effects_.bsitar}
\alias{conditional_effects_.bsitar}
\alias{conditional_effects_}
\title{Visualize the conditional effects of predictor}
\usage{
\method{conditional_effects_}{bsitar}(
  model,
  resp = NULL,
  deriv = 0,
  envir = parent.frame(),
  ...
)

conditional_effects_(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar} function for details).}

\item{deriv}{An integer value to specify whether to estimate distance curve
(i.e., model estimated curve(s)) or velocity curve (first derivative of the
model estimated curve(s)). A 0 value (default) is for distance curve and
1 for the velocity curve.}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}}
function. Please see \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}} for details.}
}
\value{
An object of class \code{'brms_conditional_effects'} which is a
  named list with one data.frame per effect containing all information
  required to generate conditional effects plots. Among others, these
  data.frames contain some special variables, namely \code{estimate__}
  (predicted values of the response), \code{se__} (standard error of the
  predicted response), \code{lower__} and \code{upper__} (lower and upper
  bounds of the uncertainty interval of the response), as well as
  \code{cond__} (used in faceting when \code{conditions} contains multiple
  rows).

  The corresponding \code{plot} method returns a named
  list of \code{\link{ggplot}} objects, which can be further
  customized using the \pkg{ggplot2} package.
}
\description{
Display conditional effects of one or more numeric and/or categorical
predictors including two-way interaction effects.
}
\details{
The \code{conditional_effects_} function is a wrapper around
the \link[brms:conditional_effects.brmsfit]{brms::conditional_effects}. The \link[brms:conditional_effects.brmsfit]{brms::conditional_effects}
function from the \code{brms} package can used to plot the fitted (distance)
curve for an \emph{bsitar} model when outcome (e.g., height) is untransformed.
However, when the outcome is log or square root transformed, the
\link[brms:conditional_effects.brmsfit]{brms::conditional_effects} will return the fitted curve on the log or
square root scale whereas the \link{conditional_effects_} will
return the fitted curve on the original scale. Furthermore, the
conditional_effects_ also displays the velocity curve on the original scale
after making required back-transformation. Apart from these differences,
both these functions (\link[brms:conditional_effects.brmsfit]{brms::conditional_effects} and
\link{conditional_effects_}) work in the same manner. In other words,
user can specify all the arguments which are available in the
\link[brms:conditional_effects.brmsfit]{brms::conditional_effects}. Because of this, the name is kept same except
for adding an underscore at the end of the name (\emph{conditional_effects} to
\emph{conditional_effects_}).
}
\examples{
\dontrun{
# The examples below show the use of *conditional_effects_* to estimate  
# the population average and individual-specific distance and velocity 
# curves.
# Population average distance curve
conditional_effects_(model, deriv = 0, re_formula = NA)

# Individual-specific distance curves
conditional_effects_(model, deriv = 0, re_formula = NULL)

# Population average velocity curve
conditional_effects_(model, deriv = 1, re_formula = NA)

# Individual-specific velocity curves
conditional_effects_(model, deriv = 1, re_formula = NULL)
 
}
}
