% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects_bgm.R
\name{conditional_effects_bgm.bgmfit}
\alias{conditional_effects_bgm.bgmfit}
\alias{conditional_effects_bgm}
\title{Visualize the conditional effects of predictor}
\usage{
\method{conditional_effects_bgm}{bgmfit}(
  model,
  resp = NULL,
  deriv = 0,
  deriv_model = TRUE,
  envir = parent.frame(),
  ...
)

conditional_effects_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}. function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar::bgm()} for details).}

\item{deriv}{An integer to specify whether to estimate distance curve or
derivatives (velocity and acceleration curves). Default \code{deriv = 0} is
for the distance curve whereas \code{deriv = 1} for velocity curve and
\code{deriv = 2} for the acceleration curve.}

\item{deriv_model}{A logical (default \code{TRUE}) to indicate whether to
estimate model based derivatives or from the differentiation of the
distance curve. When model is fit with \code{decomp = 'QR'}, the only
approach available to estimate derivatives is the  differentiation of the
distance curve.}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}}
function. Please see \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}} for details.}
}
\value{
An object of class 'brms_conditional_effects' which is a named list
with one data.frame per effect containing all information required to
generate conditional effects plots. See brms::conditional_effects for details.
}
\description{
Display conditional effects of one or more numeric and/or categorical
predictors including two-way interaction effects.
}
\details{
The \code{conditional_effects_bgm} function is a wrapper around
the \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}}. The \code{\link[brms:conditional_effects.brmsfit]{brms::conditional_effects()}}
function from the \code{brms} package can used to plot the fitted (distance)
curve for an \emph{bgmfit} model when response (e.g., height) is not transformed.
However, when the outcome is log or square root transformed, the
\link[brms:conditional_effects.brmsfit]{brms::conditional_effects} will return the fitted curve on the log or
square root scale whereas the \code{\link[=conditional_effects_bgm]{conditional_effects_bgm()}} will
return the fitted curve on the original scale. Furthermore, the
conditional_effects_bgm also displays the velocity curve on the original scale
after making required back-transformation. Apart from these differences,
both these functions (\link[brms:conditional_effects.brmsfit]{brms::conditional_effects} and
\code{\link[=conditional_effects_bgm]{conditional_effects_bgm()}}) work in the same manner. In other words,
user can specify all the arguments which are available in the
\link[brms:conditional_effects.brmsfit]{brms::conditional_effects}. Because of this, the name is kept same except
for adding an underscore at the end of the name (\emph{conditional_effects} to
\emph{conditional_effects_bgm}).
}
\examples{
\dontrun{
# The examples below show the use of *conditional_effects_bgm* to estimate  
# the population average and individual-specific distance and velocity 
# curves.
# Population average distance curve
conditional_effects_bgm(model, deriv = 0, re_formula = NA)

# Individual-specific distance curves
conditional_effects_bgm(model, deriv = 0, re_formula = NULL)

# Population average velocity curve
conditional_effects_bgm(model, deriv = 1, re_formula = NA)

# Individual-specific velocity curves
conditional_effects_bgm(model, deriv = 1, re_formula = NULL)
 
}
}
