% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_bsitar_functions.R
\name{expose_bsitar_functions}
\alias{expose_bsitar_functions}
\title{Expose Stan function for post-processing of posterior sample obtained
from the Bayesian SITAR growth curve model fit}
\usage{
expose_bsitar_functions(
  model,
  scode = NULL,
  expose = FALSE,
  select_model = NULL
)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{scode}{An option argument specifying the code with the user-defined
Stan function.}

\item{expose}{A logical (default \code{TRUE}) to indicate whether to expose
functions and add to the global environment.}

\item{select_model}{A string (default \code{NULL}) to indicate the model.}
}
\value{
An object of class \code{brmsfit, bsiatr} with exposed
user-defined Stan functions (when \code{expose=TRUE}).
}
\description{
Expose Stan function for post-processing of posterior sample obtained
from the Bayesian SITAR growth curve model fit
}
\examples{
\dontrun{
# 
expose_bsitar_functions(model)
#
}
}
