% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_.bsitar.R
\name{fitted_.bsitar}
\alias{fitted_.bsitar}
\alias{fitted_}
\title{Fitted (expected) values from the posterior predictive distribution}
\usage{
\method{fitted_}{bsitar}(model, resp = NULL, deriv = 0, envir = parent.frame(), ...)

fitted_(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar} function for details).}

\item{deriv}{An integer value to specify whether to estimate distance curve
(i.e., model estimated curve(s)) or velocity curve (first derivative of the
model estimated curve(s)). A 0 value (default) is for distance curve and
1 for the velocity curve.}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:fitted.brmsfit]{brms::fitted.brmsfit}
function. Please see \code{brms::fitted.brmsfit} for details on
various options available.}
}
\value{
An array of predicted mean response values. See \link[brms:fitted.brmsfit]{brms::fitted.brmsfit}
for details.
}
\description{
This method is an alias of \code{\link[brms]{posterior_epred.brmsfit}}
with additional arguments for obtaining summaries of the computed draws.
}
\details{
The \code{fitted_} function is a wrapper around
the \link[brms:fitted.brmsfit]{brms::fitted.brmsfit}. The \link[brms:fitted.brmsfit]{brms::fitted.brmsfit}
function from the \code{brms} package can used to plot the fitted (distance)
curve for an \emph{bsitar} model when outcome (e.g., height) is untransformed.
However, when the outcome is log or square root transformed, the
\link[brms:fitted.brmsfit]{brms::fitted.brmsfit} will return the fitted curve on the log or
square root scale whereas the \link{fitted_} will
return the fitted curve on the original scale. Furthermore, the
fitted_ also displays the velocity curve on the original scale
after making required back-transformation. Apart from these differences,
both these functions (\link[brms:fitted.brmsfit]{brms::fitted.brmsfit} and
\link{fitted_}) work in the same manner. In other words,
user can specify all the arguments which are available in the
\link[brms:fitted.brmsfit]{brms::fitted.brmsfit}. Because of this, the name is kept same except
for adding an underscore at the end of the name (\emph{fitted} to
\emph{fitted_}).
}
\examples{
\dontrun{
# The following examples show the use of *fitted_* to estimate  
# population average and individual-specific distance and velocity 
# curves.
# Population average distance curve
fitted_(model, deriv = 0, re_formula = NA)

# Individual-specific distance curves
fitted_(model, deriv = 0, re_formula = NULL)

# Population average velocity curve
fitted_(model, deriv = 1, re_formula = NA)

# Individual-specific velocity curves
fitted_(model, deriv = 1, re_formula = NULL)
 
}
}
