% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthparameters_predictions.R
\name{growthparameters_predictions.bgmfit}
\alias{growthparameters_predictions.bgmfit}
\alias{growthparameters_predictions}
\title{Compute growth parameters for the \code{bgmfit} model}
\usage{
\method{growthparameters_predictions}{bgmfit}(
  model,
  resp = NULL,
  ndraws = NULL,
  draw_ids = NULL,
  newdata = NULL,
  re_formula = NA,
  at_age = NULL,
  tsv = NULL,
  parameter = NULL,
  xrange = 1,
  acg_velocity = 0.1,
  digits = 2,
  levels_id = NULL,
  ipts = NULL,
  seed = 123,
  future = FALSE,
  future_session = "multisession",
  cores = NULL,
  average = FALSE,
  variables = NULL,
  type = NULL,
  by = FALSE,
  conf_level = 0.95,
  transform = NULL,
  wts = NULL,
  hypothesis = NULL,
  equivalence = NULL,
  reformat = TRUE,
  envir = parent.frame(),
  ...
)

growthparameters_predictions(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{resp}{An optional character string to specify response variable when
estimating growth parameter for the univariate-by-subgroup and multivariate
models (see \code{\link[=bgm]{bgm()}} for details on univariate-by-subgroup and
multivariate models). For univariate model, \code{resp = NULL} (default).}

\item{ndraws}{Positive integer indicating the number of posterior draws to be
used in estimation. If \code{NULL} (default), all draws are used.}

\item{draw_ids}{An integer indicating the specif posterior draw(s)
to be used. If \code{NULL} (default), all draws are used.}

\item{newdata}{An optional data frame to be used in predictions. If
\code{NULL} (default), the original data from the fitted model is used.}

\item{re_formula}{Option to indicate whether or not to include the
individual/group-level effects in the estimation. When \code{NA} (default),
the individual-level effects are excluded and therefore population average
growth parameters are computed. When \code{NULL}, individual-level effects
are included in the computation and hence the growth parameters estimates
returned are individual-specific. In both situations, (i.e,, \code{NA} or
\code{NULL}), continuous and factor covariate(s) are appropriately included
in the estimation. The continuous covariates by default are set to their
means (see \code{numeric_cov_at} for details) whereas factor covariates are
left unaltered thereby allowing estimation of (factor) covariate specific
population average and individual-specific growth parameter at mean value
of continous covaristes(s).}

\item{at_age}{An integer to indicate the age at which size, (distance) and
velocity to be calculated. For the \code{'logistic3'} model, the possible
\code{at_age} are \code{1} (first maximum growth velocity), \code{2} (at
childhood growth  spurt), and \code{3} (pubertal growth spurt). For
\code{at_age = NULL} (default), parameters are calculated at the pubertal
growth spurt, i.e., \code{at_age = 3}.}

\item{tsv}{An integer to indicate whether to get the timing, size, (distance)
at the given timing or velocity at give timing. If \code{tsv = NULL}
(default), timing is returned otherwise distance if \code{tsv = 0}, or
velocity if \code{tsv = 1}.}

\item{parameter}{A single character string, or a vector of character strings
indicating the parameter to be estimated. Options available are \code{ato},
\code{vto}, \code{vpv}, \code{apv}, \code{vcg}, and \code{acg}. For
\code{parameter = NULL} (default), age at peak growth velocity is estimated.}

\item{xrange}{An integer to set the predictor range (i.e., age) when
executing the interpolation via \code{ipts}. The default \code{NULL} sets
the individual specific predictor range whereas code \code{xrange = 1} sets
same range for all individuals within the higher order grouping variable
(e.g., study). Code \code{xrange  = 2} sets the identical range dplyr::across the
entire sample. Lastly, a paired numeric values can be supplied e.g.,
\code{xrange = c(6, 20)} will set the range between 6 and 20.}

\item{acg_velocity}{A real number to set the percentage of peak growth growth
velocity as the cessation velocity when estimating the \code{vcg} and
\code{acg} growth parameters. The \code{acg_velocity} should be greater
than \code{0} and less than \code{1}. The default \code{acg_velocity =
  0.10} indicates that a 10 per cent of the peak growth velocity will be used
to get the cessation velocity and the corresponding age at the cessation
velocity. For example if peak growth velocity estimate is \code{10
  mm/year}, then cessation velocity is \code{1 mm/year}.}

\item{digits}{An integer (default \code{2}) to set the decimal places for the
estimated growth parameters. The \code{digits} is passed on to the
\code{\link[base:Round]{base::round()}} function.}

\item{levels_id}{An optional argument to specify the ids for hierarchical
model (default \code{NULL}. It is used only when model is fitted to the
data with 3 or more levels of hierarchy. For a two level model, the id for
second level is automatically inferred from the fitted model. Even for 3 or
higher level model, ids are inferred from the fitted model but under the
assumption that hierarchy is specified from lower to upper levels i.e, id,
study assuming that id is nested within the studies. However, it is not
gauranted that these ids are sorted correctly. Therefore, it is better to
set them manually.}

\item{ipts}{An integer to set the length of the predictor variable to get a
smooth velocity curve. The \code{NULL} (default) will return original
values whereas an integer (e.g., \code{ipts = 10}) will interpolate the
predictor. It is important to note that these interpolations do not alter
the range of predictor when calculating population average and the
individual specific velocity curves.}

\item{seed}{An integer (default \code{123}) that is passed to the estimation
method.}

\item{future}{A logical (default \code{FALSE}) to specify whether or not to
perform parallel computations. If set to \code{TRUE}, the
\code{\link[future.apply:future_lapply]{future.apply::future_sapply()}} function is used for summarizing the draws.}

\item{future_session}{A character string to set the session type when
\code{future = TRUE}. The \code{multisession} (default) options sets the
multisession whereas the \code{multicore} sets the multicore session. Note
that multicore session is not supported on Windows systems. For more
details, see \code{\link[future.apply:future_lapply]{future.apply::future_sapply()}}.}

\item{cores}{Number of cores to be used when running the parallel
computations by setting the option \code{future = TRUE}. On non-Windows,
systems this argument can be set globally via the mc.cores option. For the
default \code{NULL} option, the number of cores are set automatically by
calling the \code{\link[future:re-exports]{future::availableCores()}}. The number of cores used are the
maximum number of cores avaialble minus one, i.e.,
\code{future::availableCores() - 1}.}

\item{average}{A logical to indicate whether to internally call the
\code{\link[marginaleffects:predictions]{marginaleffects::predictions()}} or the
\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}} function. If \code{FALSE} (default),
\code{\link[marginaleffects:predictions]{marginaleffects::predictions()}} is called otherwise
\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}} when \code{average = TRUE}.}

\item{variables}{For estimating growth parameters in the current use case,
the \code{variables} is the level 1 predictor such as \code{age}/\code{time}
and is is a named list with value of list set via the \code{esp}
(default 1e-6). If \code{NULL}, the \code{variables} is set internally by
retrieving the relevant information from the \code{model}. Otherwise,
can define it as \code{variables = list('age' = 1e-6)}.}

\item{type}{string indicates the type (scale) of the predictions used to
compute contrasts or slopes. This can differ based on the model
type, but will typically be a string such as: "response", "link", "probs",
or "zero". When an unsupported string is entered, the model-specific list of
acceptable values is returned in an error message. When \code{type} is \code{NULL}, the
first entry in the error message is used by default.}

\item{by}{Aggregate unit-level estimates (aka, marginalize, average over). Valid inputs:
\itemize{
\item \code{FALSE}: return the original unit-level estimates.
\item \code{TRUE}: aggregate estimates for each term.
\item Character vector of column names in \code{newdata} or in the data frame produced by calling the function without the \code{by} argument.
\item Data frame with a \code{by} column of group labels, and merging columns shared by \code{newdata} or the data frame produced by calling the same function without the \code{by} argument.
\item See examples below.
\item For more complex aggregations, you can use the \code{FUN} argument of the \code{hypotheses()} function. See that function's documentation and the Hypothesis Test vignettes on the \code{marginaleffects} website.
}}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval.}

\item{transform}{string or function. Transformation applied to unit-level estimates and confidence intervals just before the function returns results. Functions must accept a vector and return a vector of the same length. Support string shortcuts: "exp", "ln"}

\item{wts}{string or numeric: weights to use when computing average contrasts or slopes. These weights only affect the averaging in \verb{avg_*()} or with the \code{by} argument, and not the unit-level estimates themselves. Internally, estimates and weights are passed to the \code{weighted.mean()} function.
\itemize{
\item string: column name of the weights variable in \code{newdata}. When supplying a column name to \code{wts}, it is recommended to supply the original data (including the weights variable) explicitly to \code{newdata}.
\item numeric: vector of length equal to the number of rows in the original data or in \code{newdata} (if supplied).
}}

\item{hypothesis}{specify a hypothesis test or custom contrast using a numeric value, vector, or matrix, a string, or a string formula.
\itemize{
\item Numeric:
\itemize{
\item Single value: the null hypothesis used in the computation of Z and p (before applying \code{transform}).
\item Vector: Weights to compute a linear combination of (custom contrast between) estimates. Length equal to the number of rows generated by the same function call, but without the \code{hypothesis} argument.
\item Matrix: Each column is a vector of weights, as describe above, used to compute a distinct linear combination of (contrast between) estimates. The column names of the matrix are used as labels in the output.
}
\item String formula to specify linear or non-linear hypothesis tests. If the \code{term} column uniquely identifies rows, terms can be used in the formula. Otherwise, use \code{b1}, \code{b2}, etc. to identify the position of each parameter. The \verb{b*} wildcard can be used to test hypotheses on all estimates. Examples:
\itemize{
\item \code{hp = drat}
\item \code{hp + drat = 12}
\item \code{b1 + b2 + b3 = 0}
\item \verb{b* / b1 = 1}
}
\item String:
\itemize{
\item "pairwise": pairwise differences between estimates in each row.
\item "reference": differences between the estimates in each row and the estimate in the first row.
\item "sequential": difference between an estimate and the estimate in the next row.
\item "revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses, as described above.
}
\item See the Examples section below and the vignette: https://marginaleffects.com/articles/hypothesis.html
}}

\item{equivalence}{Numeric vector of length 2: bounds used for the two-one-sided test (TOST) of equivalence, and for the non-inferiority and non-superiority tests. See Details section below.}

\item{reformat}{A logical (default \code{TRUE}) to reformat returned output
as a data.frame with colnames renamed as follows: \code{estimate} as
\code{Estimate}, \code{conf.low} as \code{Q2.5}, and \code{conf.high} as
\code{Q97.5} (assuming that \code{conf_int = 0.95}). Also, following
columns are dropped from the data frame: \code{term}, \code{contrast},
\code{tmp_idx}, \code{predicted_lo}, \code{predicted_hi}, \code{predicted}.}

\item{envir}{Indicator to set the environment of function evaluation.
The default is \code{parent.frame}.}

\item{...}{Further arguments passed to \code{brms::fitted()} and
\code{brms::predict()} functions that control several aspects of data
validation and prediction. See \code{\link[brms:fitted.brmsfit]{brms::fitted.brmsfit()}} and
\code{\link[brms:predict.brmsfit]{brms::predict.brmsfit()}} for details.}
}
\value{
A data frame objects with estimates and CIs for computed parameter(s)
}
\description{
The \code{growthparameters_predictions} function provides
estimates of growth parameters (such as peak velocity and the age at peak
velocity) derived from the growth velocity curve. This function is a wrapper
around the \code{\link[marginaleffects:predictions]{marginaleffects::predictions()}} and
\code{\link[marginaleffects:predictions]{marginaleffects::avg_predictions()}}. A detailed explanation is available at
\url{https://marginaleffects.com/articles/predictions.html} and
\url{https://marginaleffects.com/}.
}
\details{
The \code{growthparameters_predictions} function estimates and
returns the following growth paramaeters:
\itemize{
\item pv  - peak velocity
\item apv - age at peak velocity
\item tv  - take off velocity
\item atv - age at take off velocity
\item cv  - cessation velocity
\item acv - age at cessation velocity
}

The take off velocity is the minimum velocity before the peak velocity and it
indicates the beginning of the pubertal growth spurt. The cessation velocity
indicates the end of the active pubertal growth spurt and is calculated as
the percentage of the peak velocity (\code{pv}). The percentage of the peak
velocity used in the calculation of the the cessation velocity (\code{'cv'})
is controlled via the \code{acg_velocity} argument. The \code{acg_velocity}
takes a real value as an input (greater than 0 and less than 1). Typically, a
10 percent of \code{pv} (i.e., \code{acg_velocity = 0.1}) is considered as a
good indicator of the cessation of the active pubertal growth spurt
\insertCite{Anna2022}{bsitar}.
}
\examples{
\dontrun{
growthparameters_predictions(model, parameter = 'apv')
}

}
\references{
\insertAllCited{}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
