% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.R
\docType{data}
\name{heights}
\alias{heights}
\title{Serial heights measurements on 12 girls}
\format{
A data frame with 124 observations on the following 4 variables:
\describe{
\item{id}{factor of subject ids (levels 1:12).}
\item{age}{vector of ages (years).}
\item{height}{vector of heights (cm).}
\item{men}{vector of ages at menarche (years), where negative values
are right censored.} }
}
\usage{
heights
}
\value{
A data frame with four columns.
}
\description{
Heights of 12 girls from the Chard Growth Study measured twice a
year between 8 and 16 years of age.
}
\details{
Data details are provided in the the \emph{sitar} package
\insertCite{R-sitar}{bsitar}. The data included in the \strong{bsitar} package is
the same as \code{heights} data used in the \emph{sitar} package.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
