% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper-1.R
\name{is.bgmfit}
\alias{is.bgmfit}
\title{Checks if object is of class \code{bgmfit}}
\usage{
is.bgmfit(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if object is of class \code{bgmfit}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
