% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_.bsitar.R
\name{loo_.bsitar}
\alias{loo_.bsitar}
\alias{loo_}
\title{Leave-one-out (loo) cross-validation for \code{bsitar} object}
\usage{
\method{loo_}{bsitar}(model, resp = NULL, envir = .GlobalEnv, ...)

loo_(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar} function for details).}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:loo.brmsfit]{brms::loo}
function. Please see \code{brms::loo} for details on
various options available.}
}
\value{
If just one object is provided, an object of class \code{loo}. If
multiple objects are provided, an object of class \code{loolist}.
}
\description{
Perform approximate leave-one-out cross-validation based
on the posterior likelihood using the \pkg{loo} package.
For more details see \code{\link[loo:loo]{loo}}.
}
\details{
The \code{loo_} function is a wrapper around
the \link[brms:loo.brmsfit]{brms::loo} and works exactly the same way as
\link{loo}.
}
\examples{
\dontrun{
loo_(model)
}
}
