% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_bgm.R
\name{loo_bgm.bgmfit}
\alias{loo_bgm.bgmfit}
\alias{loo_bgm}
\title{Leave-one-out (loo) cross-validation for \code{bgmfit} object}
\usage{
\method{loo_bgm}{bgmfit}(
  model,
  resp = NULL,
  cores = 1,
  envir = parent.frame(),
  deriv = NULL,
  ...
)

loo_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bgmfit} function for details).}

\item{cores}{Number of cores to be used when running the parallel
computations by setting the option \code{future = TRUE}. On non-Windows,
systems this argument can be set globally via the mc.cores option. For the
default \code{NULL} option, the number of cores are set automatically by
calling the \code{\link[future:re-exports]{future::availableCores()}}. The number of cores used are the
maximum number of cores avaialble minus one, i.e.,
\code{future::availableCores() - 1}.}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{deriv}{Must be \code{NULL}.}

\item{...}{Additional arguments passed to the \link[brms:loo.brmsfit]{brms::loo}
function. Please see \code{brms::loo} for details on
various options available.}
}
\value{
If just one object is provided, an object of class \code{loo}. If
multiple objects are provided, an object of class \code{loolist}.
}
\description{
Perform approximate leave-one-out cross-validation based
on the posterior likelihood using the \pkg{loo} package.
For more details see \code{\link[loo:loo]{loo}}.
}
\details{
The \code{loo_bgm} function is a wrapper around
the \link[brms:loo.brmsfit]{brms::loo} and works exactly the same way as
\link[brms:loo.brmsfit]{brms::loo}.
}
\examples{
\dontrun{
loo_bgm(model)
}
}
