% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize.bsitar}
\alias{optimize.bsitar}
\title{optimize \pkg{bsitar} models}
\usage{
\method{optimize}{bsitar}(
  model,
  newdata = NULL,
  recompile = T,
  optimize = list(x = c("NULL", "log", "sqrt"), y = c("NULL", "log", "sqrt")),
  add_fit_criteria = NULL,
  add_fit_bayes_R = NULL,
  ...
)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{newdata}{newdata Optional \code{data.frame} to update the model with
new data. Note that data-dependent default priors will not be updated
automatically.}

\item{recompile}{A logical to indicate whether the Stan model should be
recompiled. When \code{NULL} (the default), \code{update} tries to figure
out internally, if recompilation is necessary. Setting it to \code{FALSE}
will cause all Stan code changing arguments to be ignored.}

\item{optimize}{A list specifying the transformations of predictor (typically
\code{age}) and the outcome.}

\item{add_fit_criteria}{An optional (default \code{NULL}) indicator to add
fit criteria to the model fit. options are \code{loo} and \code{waic}.
Please see \code{\link[brms:add_criterion]{brms::add_criterion()}} for details.}

\item{add_fit_bayes_R}{An optional (default \code{NULL}) to add Bayesian R
square.}

\item{...}{Other arguments passed to \code{\link{brms}}.}
}
\value{
An updated object of class \code{brmsfit, bsiatr}, that contains the
posterior draws and other useful information about the model.
}
\description{
optimize \pkg{bsitar} models
}
\examples{
\dontrun{
data(heights)
data_males <- heights \%>\% filter(sex == 'Male')
fit_males <- bsitar(x=age, y=height, id=id, data=heights, df=4)
fit_males2 <- update(df=5)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
