% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_bgm.R
\name{optimize_bgm.bgmfit}
\alias{optimize_bgm.bgmfit}
\alias{optimize_bgm}
\title{Optimize \pkg{bgmfit} model}
\usage{
\method{optimize_bgm}{bgmfit}(
  model,
  newdata = NULL,
  optimize_df = NULL,
  optimize_x = list(NULL, log, sqrt),
  optimize_y = list(NULL, log, sqrt),
  exclude_default_funs = TRUE,
  add_fit_criteria = NULL,
  add_fit_bayes_R = NULL,
  byresp = FALSE,
  digits = 2,
  ...
)

optimize_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{newdata}{An optional \code{data.frame} to be used when optimizing the
model. If \code{NULL} (default), the same same data used for original model
fit is used. Note that data-dependent default priors will not be updated
automatically.}

\item{optimize_df}{A vector specifying the degree of freedom (\code{df})
update. If \code{NULL} (default), the \code{df} is taken from the
original model. For \code{univariate-by-sungroup} and \code{multivariate}
models (see \code{\link[=bgm]{bgm()}} for details on these arguments),
\code{optimize_df} can be a single integer (e.g., \code{optimize_df = 4}) or
a list (e.g., \code{optimize_df = list(4,5)}). For optimization over
different \code{df}, say for example df 4 and 5 for univariate model, the
corresponding code is \code{optimize_df = list(4,5)}. For a multivariate
model fit to two outcomes with different \code{df}, the optimization over
\code{df} 4 and 5 for the first sub model and 5 and 6 for the second
sub model, the corresponding \code{optimize_df} code is \code{optimize_df =
 list(list(4,5), list(5,6))} i.e, a list of lists.}

\item{optimize_x}{A vector specifying the transformations of predictor
(typically \code{age}) variable (via \code{xvar}). The option are
\code{NULL}, \code{log}, \code{sqrt} and their combinations. Note that user
need not to enclose these options in a single or double quotes as they are
take care of internally. The default setting is to explore all possible
combination i.e., \code{optimize_x = list(NULL, log,  sqrt)}. Similar to
the \code{optimize_df}, user can specify different \code{optimize_x} for
\code{univariate-by-sungroup} and \code{multivariate} sub models.}

\item{optimize_y}{A vector specifying the transformations for the response
variable (via \code{yvar}). The approach and options available for
\code{optimize_y} are identical to the \code{optimize_x} (see above).}

\item{exclude_default_funs}{A logical to indicate whether transformations for
(\code{xvar} and \code{yvar}) used in the original model fit should be
excluded. If \code{TRUE} (default), the \code{xvar} and \code{yvar}
transformations specified for the original model fit are excluded from the
\code{optimize_x} and \code{optimize_y}. From example, if original model is
fit with \code{xvar = log} and \code{yvar = NULL}, then \code{optimize_x} is
translated into \code{optimize_x = list(NULL, sqrt)} and \code{optimize_y}
as \code{optimize_y = list(log, sqrt)}.}

\item{add_fit_criteria}{An optional (default \code{NULL}) indicator to add fit
criteria to the model fit. options are \code{loo} and \code{waic}. Please
see \code{\link[brms:add_criterion]{brms::add_criterion()}} for details.}

\item{add_fit_bayes_R}{An optional (default \code{NULL}) to add Bayesian R
square.}

\item{byresp}{A logical (default \code{FALSE}) to indicate if response wise
fit criteria to be calculated. This argument is evaluated only for the
\code{multivariate} model for which options are available for joint
calculation of pointwise log likelihood or response specific. For,
\code{univariate-by-subgroup} model, the only option available is to
calculate separate pointwise log likelihood for each sub-model.}

\item{digits}{An integer to set the number of decimal places.}

\item{...}{Other arguments passed to \code{\link{update_bgm}}.}
}
\value{
A list containing the optimized models of class \code{brmsfit,
 bsiatr}, and the the combined summary statistics if \code{add_fit_criteria}
and/or \code{add_fit_bayes_R} are specified.
}
\description{
Optimize \pkg{bgmfit} model
}
\examples{
\dontrun{
data(heights)
data_males <- heights \%>\% filter(sex == 'Male')
fit_males <- bgm(x=age, y=height, id=id, data=heights, df=4)
fit_males2 <- optimize_bgm(fit_males)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
