% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_model.R
\name{optimize_model.bgmfit}
\alias{optimize_model.bgmfit}
\alias{optimize_model}
\title{Optimize model}
\usage{
\method{optimize_model}{bgmfit}(
  model,
  newdata = NULL,
  optimize_df = NULL,
  optimize_x = list(NULL, log, sqrt),
  optimize_y = list(NULL, log, sqrt),
  exclude_default_funs = TRUE,
  add_fit_criteria = NULL,
  add_bayes_R = NULL,
  byresp = FALSE,
  digits = 2,
  cores = 1,
  verbose = FALSE,
  envir = NULL,
  ...
)

optimize_model(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{newdata}{An optional \code{data.frame} to be used when optimizing the
model. If \code{NULL} (default), the same same data used for original model
fit is used. Note that data-dependent default priors will not be updated
automatically.}

\item{optimize_df}{A vector of integers specifying the degree of freedom
(\code{df}) values to be updated. If \code{NULL} (default), the \code{df} is
taken from the original model. For \code{univariate-by-sungroup} and
\code{multivariate} models (see \code{\link[=bsitar]{bsitar()}} for details),
\code{optimize_df} can be a single integer (e.g., \code{optimize_df = 4}) or
a list (e.g., \code{optimize_df = list(4,5)}). For optimization over
different \code{df}, say for example \code{df} 4 and \code{df} 5 for
univariate model, the corresponding code is \code{optimize_df = list(4,5)}.
For a multivariate model fit to two outcomes with different \code{df}, the
optimization over \code{df} 4 and \code{df} 5 for the first sub model, and
\code{df} 5 and \code{df} 6 for the second
sub model, the corresponding \code{optimize_df} code is \code{optimize_df =
 list(list(4,5), list(5,6))} i.e, a list of lists.}

\item{optimize_x}{A vector specifying the transformations of predictor
variable (i.e., \code{xvar} which is typically \code{age}). The option are
\code{NULL}, \code{log}, \code{sqrt}, and their combinations. Note that user
need not to enclose these options in a single or double quotes as they are
take care of internally. The default setting is to explore all possible
combination i.e., \code{optimize_x = list(NULL, log,  sqrt)}. Similar to the
\code{optimize_df}, user can specify different \code{optimize_x} for
\code{univariate-by-sungroup} and \code{multivariate} sub models.}

\item{optimize_y}{A vector specifying the transformations of the response
variable (i.e., \code{yvar} which is typically repeated height
measurements). The approach and options available for \code{optimize_y} are
identical to those described above for the \code{optimize_x}.}

\item{exclude_default_funs}{A logical to indicate whether transformations for
(\code{xvar} and \code{yvar}) used in the original model fit should be
excluded. If \code{TRUE} (default), the \code{xvar} and \code{yvar}
transformations specified for the original model fit are excluded from the
\code{optimize_x} and \code{optimize_y}. From example, if original model is
fit with \code{xvar = log} and \code{yvar = NULL}, then \code{optimize_x} is
translated into \code{optimize_x = list(NULL, sqrt)} and \code{optimize_y}
as \code{optimize_y = list(log, sqrt)}.}

\item{add_fit_criteria}{An optional (default \code{NULL}) indicator to add fit
criteria to returned model fit. Options are \code{loo} and \code{waic}.
Please see \code{\link[brms:add_criterion]{brms::add_criterion()}} for details.}

\item{add_bayes_R}{An optional (default \code{NULL}) to add Bayesian R
square. To estimate \code{bayes_R2}, please use
\code{add_bayes_R = bayes_R2}}

\item{byresp}{A logical (default \code{FALSE}) to indicate if response wise
fit criteria to be calculated. This argument is evaluated only for the
\code{multivariate} model for which user can select whether to get joint
calculation of point wise log likelihood or response specific. For,
\code{univariate-by-subgroup} model, the only option available is to
calculate separate point wise log likelihood for each sub-model.}

\item{digits}{An integer to set the number of decimal places.}

\item{cores}{The number of cores to used in processing (default \code{1}).
This passed to the \code{\link[brms:add_criterion]{brms::add_criterion()}}.}

\item{verbose}{A logical (default \code{FALSE}) to indicate whether to print
information.}

\item{envir}{Environment used for function evaluation. The default is
\code{NULL} which will set \code{parent.frame()} as default environment.
Note that since most of post processing functions are based on \pkg{brms},
the functions needed for evaluation should be in the \code{.GlobalEnv}.
Therefore, it is strongly recommended to set \code{ envir = globalenv()}
(or \code{envir = .GlobalEnv}). This is particularly true for the
derivatives such as velocity curve.}

\item{...}{Other arguments passed to \code{\link{update_model}}.}
}
\value{
A list containing the optimized models of class \code{bgmfit}, and the
the summary statistics if \code{add_fit_criteria} and/or
\code{add_bayes_R} are specified.
}
\description{
Optimize model
}
\examples{

# Fit Bayesian SITAR model 

# To avoid fitting the model which takes time, the model  
# fit has already been saved as 'berkeley_mfit.rda' file.
# See examples section of the main function for details on the model fit.

model <- berkeley_mfit

\donttest{

# Fit Bayesian SITAR model 

# To avoid mode estimation which takes time, a model fitted to the 
# 'berkeley_mdata' has already been saved as 'berkeley_mfit'. Details
# on 'berkeley_mdata' and 'berkeley_mfit' are provided in 'bsitar' function.

model <- berkeley_mfit

# Below example shows optimization of the degree of freedom (df). 
# Note that in case degree of freedom is same as the original model, and both 
# optimize_x and optimize_y NULL (i.e., nothing to optimize), then original  
# model is returned without evaluating any other argument. Since nothing is 
# set to  be optimized in the below example, the original model is returned.  
# To explicitly get this information whether model is being optimized or not, 
# user can set verbose = TRUE. The verbose = TRUE also useful in getting the
# information regarding what all arguments have been changed as compared to
# the original model.

model2 <- optimize_model(model, optimize_df = 4, 
  optimize_x = NULL, 
  optimize_y = NULL,
  verbose = TRUE)

}

}
\seealso{
\code{\link[brms:add_criterion]{brms::add_criterion()}}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
