% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Identify (and remove) outliers with abnormal growth velocity}
\usage{
outliers(
  x,
  y,
  id,
  data,
  icode = c(4:6),
  lag = 1,
  velpower = 0.5,
  limit = 5,
  linearise = FALSE,
  remove = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{age vector.}

\item{y}{outcome vector, typically weight or height.}

\item{id}{factor identifying each subject.}

\item{data}{data frame containing x, y and id.}

\item{icode}{The code number(s) defining the subset of curves to be
displayed or zapped (between 1 and 6).}

\item{lag}{lag between measurements for defining growth velocity.}

\item{velpower}{a value, typically between 0 and 1, defining the power of
delta x to use when calculating velocity as delta(y)/delta(x)^velpower. The
default of 0.5 is midway between velocity and increment.}

\item{limit}{the number of standard deviations beyond which a velocity is
deemed to be an outlier.}

\item{linearise}{if TRUE y is converted to a residual about the median curve
of y versus x.}

\item{remove}{A logical (default \code{FALSE}) to indicate whether identified
remove to be removed. When \code{FALSE}, outliers are set as \code{NA}. If
\code{TRUE}, outliers set as \code{NA} are removed.}

\item{verbose}{A logical (default \code{FALSE}) to show frequency of
observations with different codes (see \code{\link[sitar:velout]{sitar::velout()}}), and the number
of observations (outliers) removed when \code{remove=TRUE}.}
}
\value{
A data frame with outliers removed when \code{remove=TRUE}
}
\description{
The function identifies and remove the putative outliers for y
in data, Codes range from 0 (normal) to 8, where 4 and 6 are conventional
outliers
}
\examples{

data <- berkeley_mdata

outliers(x = "age", y = "height", id = "id", data = data, remove=TRUE)

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
