% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.R
\name{plot_curves.bgmfit}
\alias{plot_curves.bgmfit}
\alias{plot_curves}
\title{Plot growth curves for the Bayesian SITAR model}
\usage{
\method{plot_curves}{bgmfit}(
  model,
  opt = "dv",
  apv = FALSE,
  bands = NULL,
  conf = 0.95,
  resp = NULL,
  dpar = NULL,
  ndraws = NULL,
  draw_ids = NULL,
  newdata = NULL,
  summary = FALSE,
  digits = 2,
  re_formula = NULL,
  numeric_cov_at = NULL,
  aux_variables = NULL,
  cov = NULL,
  grid_call = NULL,
  grid_by = NULL,
  grid_add = NULL,
  levels_id = NULL,
  avg_reffects = NULL,
  ipts = NULL,
  model_deriv = NULL,
  xrange = NULL,
  xrange_search = NULL,
  takeoff = FALSE,
  trough = FALSE,
  acgv = FALSE,
  acgv_velocity = 0.1,
  seed = 123,
  estimation_method = "fitted",
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty",
  incl_autocor = TRUE,
  robust = FALSE,
  transform_draws = NULL,
  transform_xaxis = list(x_min = NULL, x_max = NULL, n = 5, by = NULL, xaxis_fun = NULL,
    spacing = "data", accuracy = 0.1, expand_mult = 0, use_coord_limits = TRUE,
    coord_pad_mult = 0.05, coord_pad_add = 0),
  scale = c("response", "linear"),
  future = FALSE,
  future_session = "multisession",
  cores = NULL,
  trim = 0,
  layout = "single",
  linecolor = NULL,
  linecolor1 = NULL,
  linecolor2 = NULL,
  label.x = NULL,
  label.y = NULL,
  label.title = NULL,
  label.subtitle = NULL,
  legendpos = "bottom",
  linetype.apv = NULL,
  linewidth.main = NULL,
  linewidth.apv = NULL,
  linetype.groupby = NA,
  color.groupby = NULL,
  fill.groupby = NULL,
  band.alpha = NULL,
  band.line = FALSE,
  band.legends = FALSE,
  mult = c(0.02, 0.05),
  expand = c(TRUE, TRUE, FALSE, FALSE),
  show_age_takeoff = TRUE,
  show_age_peak = TRUE,
  show_age_cessation = TRUE,
  show_vel_takeoff = FALSE,
  show_vel_peak = FALSE,
  show_vel_cessation = FALSE,
  each_object = FALSE,
  ncol = NULL,
  print = TRUE,
  returndata = FALSE,
  returndata_add_parms = FALSE,
  parms_eval = FALSE,
  idata_method = NULL,
  parms_method = "getPeak",
  verbose = FALSE,
  fullframe = NULL,
  dummy_to_factor = NULL,
  expose_function = FALSE,
  usesavedfuns = NULL,
  clearenvfuns = NULL,
  funlist = NULL,
  xvar = NULL,
  difx = NULL,
  idvar = NULL,
  itransform = NULL,
  newdata_fixed = NULL,
  envir = NULL,
  ...
)

plot_curves(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{opt}{A character string containing one or more of the following
plotting options:
\itemize{
\item 'd': Population average distance curve
\item 'v': Population average velocity curve
\item 'D': Individual-specific distance curves
\item 'V': Individual-specific velocity curves
\item 'u': Unadjusted individual-specific distance curves
\item 'a': Adjusted individual-specific distance curves (adjusted for
random effects)
}}

\item{apv}{A logical value (default \code{FALSE}) indicating whether to
calculate and plot the age at peak velocity (APGV) when \code{opt} includes
'v' or 'V'.}

\item{bands}{A character string containing one or more of the following
options, or \code{NULL} (default), indicating if CI bands should be plotted
around the curves:
\itemize{
\item 'd': Band around the distance curve
\item 'v': Band around the velocity curve
\item 'p': Band around the vertical line denoting the APGV parameter
}

The \code{'dvp'} option will include CI bands for distance and velocity
curves, and the APGV.}

\item{conf}{A numeric value (default \code{0.95}) specifying the confidence
interval (CI) level for the bands. See \code{\link[=growthparameters]{growthparameters()}} for
more details.}

\item{resp}{A character string (default \code{NULL}) to specify the response
variable when processing posterior draws for \code{univariate_by} and
\code{multivariate} models. See \code{\link[=bsitar]{bsitar()}} for details on
\code{univariate_by} and \code{multivariate} models.}

\item{dpar}{Optional name of a predicted distributional parameter.
If specified, expected predictions of this parameters are returned.}

\item{ndraws}{A positive integer indicating the number of posterior draws to
use in estimation. If \code{NULL} (default), all draws are used.}

\item{draw_ids}{An integer specifying the specific posterior draw(s) to use
in estimation (default \code{NULL}).}

\item{newdata}{An optional data frame for estimation. If \code{NULL}
(default), \code{newdata} is retrieved from the \code{model}.}

\item{summary}{A logical value indicating whether only the estimate should be
computed (\code{TRUE}), or whether the estimate along with SE and CI should
be returned (\code{FALSE}, default). Setting \code{summary} to \code{FALSE}
will increase computation time. Note that \code{summary = FALSE} is
required to obtain correct estimates when \code{re_formula = NULL}.}

\item{digits}{An integer (default \code{2}) to set the decimal places for
rounding the results using the \code{\link[base:round]{base::round()}} function.}

\item{re_formula}{Option to indicate whether or not to include
individual/group-level effects in the estimation. When \code{NA} (default),
individual-level effects are excluded, and population average growth
parameters are computed. When \code{NULL}, individual-level effects are
included in the computation, and the resulting growth parameters are
individual-specific. In both cases (\code{NA} or \code{NULL}), continuous
and factor covariates are appropriately included in the estimation.
Continuous covariates are set to their means by default (see
\code{numeric_cov_at} for details), while factor covariates remain
unaltered, allowing for the estimation of covariate-specific population
average and individual-specific growth parameters.}

\item{numeric_cov_at}{An optional (named list) argument to specify the value
of continuous covariate(s). The default \code{NULL} option sets the
continuous covariate(s) to their mean. Alternatively, a named list can be
supplied to manually set these values. For example, \code{numeric_cov_at =
list(xx = 2)} will set the continuous covariate variable 'xx' to 2. The
argument \code{numeric_cov_at} is ignored when no continuous covariates are
included in the model.}

\item{aux_variables}{An optional argument to specify variables passed to the
\code{ipts} argument, useful when fitting location-scale or measurement
error models.}

\item{cov}{An optional character string to specify the \code{covariate}
passed to the plot argument (default  \code{NULL}). Rarely used.}

\item{grid_call}{An optional logical (default  \code{NULL}) to indicate
whether to call \code{\link[marginaleffects:datagrid]{marginaleffects::datagrid()}} when \code{dpar = 'sigma'}.
Rarely used. The two indications are when \code{sigma_model = 'ls'} or
\code{opt = 'v'}.}

\item{grid_by}{An optional character string to specify the \code{by} argument
passed to the \code{\link[marginaleffects:datagrid]{marginaleffects::datagrid()}} (default  \code{NULL}). Rarely
used. Ignored when \code{grid_call = FALSE}. If \code{grid_call = TRUE} and
\code{grid_by = NULL}, then internally set as \code{grid_by = cov}.}

\item{grid_add}{An optional argument to specify the variable(s) that can be
passed to the \code{\link[marginaleffects:datagrid]{marginaleffects::datagrid()}}. This is useful when fitting
location-scale models and measurement error models. If
post-processing functions throw an error such as \code{variable 'x' not
found in either 'data' or 'data2'}, consider using \code{grid_add}.
Note that unlike \code{aux_variables} which are passed to the internal data
functions such as \code{'get.newdata'}, the \code{grid_add} are passed to
the \code{\link[marginaleffects:datagrid]{marginaleffects::datagrid()}}.}

\item{levels_id}{An optional argument to specify the \code{ids} for the
hierarchical model (default \code{NULL}). It is used only when the model is
applied to data with three or more levels of hierarchy. For a two-level
model, \code{levels_id} is automatically inferred from the model fit. For
models with three or more levels, \code{levels_id} is inferred from the
model fit under the assumption that hierarchy is specified from the lowest
to the uppermost level, i.e., \code{id} followed by \code{study}, where
\code{id} is nested within \code{study}. However, it is not guaranteed that
\code{levels_id} is sorted correctly, so it is better to set it manually
when fitting a model with three or more levels of hierarchy.}

\item{avg_reffects}{An optional argument (default \code{NULL}) to calculate
(marginal/average) curves and growth parameters, such as APGV and PGV. If
specified, it must be a named list indicating the \code{over} (typically a
level 1 predictor, such as age), \code{feby} (fixed effects, typically a
factor variable), and \code{reby} (typically \code{NULL}, indicating that
parameters are integrated over the random effects). For example,
\code{avg_reffects = list(feby = 'study', reby = NULL, over = 'age')}.}

\item{ipts}{An integer specifying the number of points for interpolating the
predictor variable (e.g., age) to generate smooth curves for predictions
and plots. This value is used as the \code{length.out} argument for
\code{\link[=seq]{seq()}}, controlling the smoothness of distance and velocity curves without
altering the predictor range.
\describe{
\item{\code{NULL} (the default)}{Engages automatic behavior based on the
\code{dpar} argument: it is internally set to \code{50} for the mean
response (\code{dpar = 'mu'}), ensuring smooth curves, while for the
distributional parameters (e.g., \code{dpar = 'sigma'}), no
interpolation is performed.}
\item{An integer (e.g., \code{100})}{Explicitly sets the number of
interpolation points.}
\item{\code{FALSE}}{Disables all interpolation, forcing predictions to
be made only at the original data points of the predictor variable.}
}

This argument affects the following post-processing functions: \cr
\code{\link[=fitted_draws]{fitted_draws()}}, \code{\link[=predict_draws]{predict_draws()}}, \code{\link[=growthparameters]{growthparameters()}}, \code{\link[=plot_curves]{plot_curves()}},
\code{\link[=get_predictions]{get_predictions()}}, \code{\link[=get_comparisons]{get_comparisons()}}, and
\code{\link[=get_growthparameters]{get_growthparameters()}}.}

\item{model_deriv}{A logical value specifying whether to estimate the
velocity curve from the derivative function or by differentiating the
distance curve. Set \code{model_deriv = TRUE} for functions that require
the velocity curve, such as \code{growthparameters()} and
\code{plot_curves()}. Set it to \code{NULL} for functions that use the
distance curve (i.e., fitted values), such as \code{loo_validation()} and
\code{plot_ppc()}.}

\item{xrange}{An integer to set the predictor range (e.g., age) when
executing the interpolation via \code{ipts}. By default, \code{NULL} sets
the individual-specific predictor range. Setting \code{xrange = 1} applies
the same range for individuals within the same higher grouping variable
(e.g., study). Setting \code{xrange = 2} applies an identical range across
the entire sample. Alternatively, a numeric vector (e.g., \code{xrange =
c(6, 20)}) can be provided to set the range within the specified values.}

\item{xrange_search}{A vector of length two or a character string
\code{'range'} to set the range of the predictor variable (\code{x}) within
which growth parameters are searched. This is useful when there is more
than one peak and the user wants to summarize the peak within a specified
range of the \code{x} variable. The default value is \code{xrange_search =
NULL}.}

\item{takeoff}{A logical value (default \code{FALSE}) indicating whether to
calculate the age at takeoff velocity (ATGV) and the takeoff growth
velocity (TGV) parameters.}

\item{trough}{A logical value (default \code{FALSE}) indicating whether to
calculate the age at cessation of growth velocity (ACGV) and the cessation
of growth velocity (CGV) parameters.}

\item{acgv}{A logical value (default \code{FALSE}) indicating whether to
calculate the age at cessation of growth velocity from the velocity curve.
If \code{TRUE}, the age at cessation of growth velocity (ACGV) and the
cessation growth velocity (CGV) are calculated based on the percentage of
the peak growth velocity, as defined by the \code{acgv_velocity} argument
(see below). The \code{acgv_velocity} is typically set at 10 percent of the
peak growth velocity. ACGV and CGV are calculated along with the
uncertainty (SE and CI) around the ACGV and CGV parameters.}

\item{acgv_velocity}{The percentage of the peak growth velocity to use when
estimating \code{acgv}. The default value is \code{0.10}, i.e., 10 percent
of the peak growth velocity.}

\item{seed}{An integer (default \code{123}) that is passed to the estimation
method to ensure reproducibility.}

\item{estimation_method}{A character string specifying the estimation method
when calculating the velocity from the posterior draws. The \code{'fitted'}
method internally calls \code{\link[=fitted_draws]{fitted_draws()}}, while the
\code{'predict'} method calls \code{\link[=predict_draws]{predict_draws()}}. See
\code{\link[brms:fitted.brmsfit]{brms::fitted.brmsfit()}} and \code{\link[brms:predict.brmsfit]{brms::predict.brmsfit()}} for details.}

\item{allow_new_levels}{A flag indicating if new levels of group-level
effects are allowed (defaults to \code{FALSE}). Only relevant if
\code{newdata} is provided.}

\item{sample_new_levels}{Indicates how to sample new levels for grouping
factors specified in \code{re_formula}. This argument is only relevant if
\code{newdata} is provided and \code{allow_new_levels} is set to
\code{TRUE}. If \code{"uncertainty"} (default), each posterior sample for a
new level is drawn from the posterior draws of a randomly chosen existing
level. Each posterior sample for a new level may be drawn from a different
existing level such that the resulting set of new posterior draws
represents the variation across existing levels. If \code{"gaussian"},
sample new levels from the (multivariate) normal distribution implied by the
group-level standard deviations and correlations. This options may be useful
for conducting Bayesian power analysis or predicting new levels in
situations where relatively few levels where observed in the old_data. If
\code{"old_levels"}, directly sample new levels from the existing levels,
where a new level is assigned all of the posterior draws of the same
(randomly chosen) existing level.}

\item{incl_autocor}{A flag indicating if correlation structures originally
specified via \code{autocor} should be included in the predictions.
Defaults to \code{TRUE}.}

\item{robust}{A logical value to specify the summary options. If \code{FALSE}
(default), the mean is used as the measure of central tendency and the
standard deviation as the measure of variability. If \code{TRUE}, the
median and median absolute deviation (MAD) are applied instead. Ignored if
\code{summary} is \code{FALSE}.}

\item{transform_draws}{A function applied to individual draws from the
posterior distribution before computing summaries (default \code{NULL}).
The argument \code{transform_draws} is derived from the
\code{\link[marginaleffects:predictions]{marginaleffects::predictions()}} function and should not be confused with
the \code{transform} argument from the deprecated
\code{\link[brms:posterior_predict]{brms::posterior_predict()}} function. It's important to note that for both
\code{\link[marginaleffects:predictions]{marginaleffects::predictions()}} and \code{\link[marginaleffects:avg_predictions]{marginaleffects::avg_predictions()}},
the \code{transform_draws} argument takes precedence over the
\code{transform} argument. Note that when \code{transform_draws = NULL}, an
attempt is made to automatically set \code{transform_draws = 'exp'} for
\code{dpar = 'sigma'}. User can set  \code{transform_draws = FALSE} to turn
off this automatic assignment of \code{'exp'} to the
\code{transform_draws}. It is also important to set \code{transform_draws =
FALSE} when computing the first derivative (velocity) for \code{dpar =
'sigma'}.}

\item{transform_xaxis}{A named \code{list()} of arguments for \code{build_scale_x()}.
Any unspecified arguments are automatically filled with the default values
from \code{build_scale_x()}.
Common elements include:
\itemize{
\item \code{x_min}: lower x-axis limit. If \code{NULL} (default),
x_min is inferred internally from the \code{xvar}
\item \code{x_max}: upper x-axis limit. If \code{NULL} (default),
x_max is inferred internally from the \code{xvar}
\item \code{n}: number of breaks.
\item \code{by}: optional step size for breaks.
\item \code{xaxis_fun}: transformation used for break placement and
inverse labeling, for example \code{"identity"}, \code{"log"}, \code{"log10"},
\code{"sqrt"}, or \code{"square"}, or a function
\item \code{spacing}: either \code{"data"} or \code{"trans"}.
\item \code{accuracy}: label accuracy.
\item \code{expand_mult}: scale expansion multiplier.
\item \code{use_coord_limits}: logical; if \code{TRUE}, zoom with
\code{ggplot2::coord_cartesian()}.
\item \code{coord_pad_mult}: multiplicative padding added to both x-axis ends.
\item \code{coord_pad_add}: additive padding added to both x-axis ends.
}}

\item{scale}{Either \code{"response"} or \code{"linear"}.
If \code{"response"}, results are returned on the scale
of the response variable. If \code{"linear"},
results are returned on the scale of the linear predictor term,
that is without applying the inverse link function or
other transformations.}

\item{future}{A logical value (default \code{FALSE}) indicating whether to
perform parallel computations. If \code{TRUE}, posterior summaries are
computed in parallel using \code{\link[future.apply:future_sapply]{future.apply::future_sapply()}}.}

\item{future_session}{A character string or a named list specifying the
parallel plan when \code{future = TRUE}.
\itemize{
\item \strong{Character string}: Defaults to \code{"multisession"}.
Use \code{"multicore"} for forking (not supported on Windows).
\item \strong{Named list}: Useful for advanced plans like
\code{\link[future.mirai:mirai_cluster]{future.mirai::mirai_cluster()}}. The list must contain:
\itemize{
\item \code{future_session}: The planner function (e.g.,
\code{mirai_cluster}).
\item Additional named arguments passed to the planner (e.g.,
\code{daemons} for \code{\link[mirai:daemons]{mirai::daemons()}}).
}
Example: \code{list(future_session = mirai_cluster, daemons =
    list(...))}.
}}

\item{cores}{An integer specifying the number of cores for parallel
execution.
\itemize{
\item If \code{NULL} (default), the number of cores is set to
\code{future::availableCores() - 1}.
\item On non-Windows systems, this can be controlled globally via the
\code{mc.cores} option.
}}

\item{trim}{A numeric value (default \code{0}) indicating the number of long
line segments to be excluded from the plot when the option 'u' or 'a' is
selected. See \link[sitar:plot.sitar]{sitar::plot.sitar} for further details.}

\item{layout}{A character string defining the plot layout. The default
\code{'single'} layout overlays distance and velocity curves on a single
plot when \code{opt} includes combinations like \code{'dv'}, \code{'Dv'},
\code{'dV'}, or \code{'DV'}. The alternative layout option \code{'facet'}
uses \code{facet_wrap} from \pkg{ggplot2} to map and draw plots when
\code{opt} includes two or more letters.}

\item{linecolor}{The color of the lines when the layout is \code{'facet'}.
The default is \code{NULL}, which sets the line color to \code{'grey50'}.}

\item{linecolor1}{The color of the first line when the layout is
\code{'single'}. For example, in \code{opt = 'dv'}, the distance line is
controlled by \code{linecolor1}. The default \code{NULL} sets
\code{linecolor1} to \code{'orange2'}.}

\item{linecolor2}{The color of the second line when the layout is
\code{'single'}. For example, in \code{opt = 'dv'}, the velocity line is
controlled by \code{linecolor2}. The default \code{NULL} sets
\code{linecolor2} to \code{'green4'}.}

\item{label.x}{An optional character string to label the x-axis. If
\code{NULL} (default), the x-axis label will be taken from the predictor
(e.g., age).}

\item{label.y}{An optional character string to label the y-axis. If
\code{NULL} (default), the y-axis label will be taken from the plot type
(e.g., distance, velocity). When \code{layout = 'facet'}, the label is
removed, and the same label is used as the title.}

\item{label.title}{An optional character string to label the title. Default
\code{NULL}.}

\item{label.subtitle}{An optional character string to label the title. Default
\code{NULL}}

\item{legendpos}{A character string to specify the position of the legend.
If, the legend position is set to 'bottom' for distance and velocity curves
in the \code{'single'} layout. For individual-specific curves, the legend
position is set to \code{'none'} to suppress the legend.}

\item{linetype.apv}{A character string to specify the type of the vertical
line marking the APGV. Default \code{NULL} sets the linetype to
\code{dotted}.}

\item{linewidth.main}{A numeric value to specify the line width for distance
and velocity curves. The default \code{NULL} sets the width to 0.35.}

\item{linewidth.apv}{A numeric value to specify the width of the vertical
line marking the APGV. The default \code{NULL} sets the width to 0.25.}

\item{linetype.groupby}{A character string specifying the line type of
distance and velocity curves for \code{covariates} or individual-specific
curves. If \code{NA}, then default color is \code{'black'} for all lines.}

\item{color.groupby}{A character string specifying the line color of distance
and velocity curves for \code{covariates} or individual-specific curves. If
\code{NA}, then default color is \code{'black'} for all lines. If
\code{NULL} (default), then line color is set based on the co variate
extracted from the \code{model_info}. Note that for individual-specific
curves for co variate adjusted models, user can set it as
\code{color.groupby = interaction(covariate, id)}}

\item{fill.groupby}{A character string specifying the fill color of distance
and velocity curves for \code{covariates} or individual-specific curves. If
\code{NA}, then default color is \code{'black'} for all lines. If
\code{NULL} (default), then fill color is set same as the
\code{color.groupby}.}

\item{band.alpha}{A numeric value to specify the transparency of the CI bands
around the curves. The default \code{NULL} sets the transparency to 0.4.}

\item{band.line}{A logical value (default \code{FALSE}) to indicate whether
to add line around the CI bands.}

\item{band.legends}{A logical value (default \code{FALSE}) to indicate
whether to show \code{legends} for \code{CI bands}. It sets
\code{ggplot2::guides(fill = 'none')}}

\item{mult}{vector of multiplicative range expansion factors.
If length 1, both the lower and upper limits of the scale
are expanded outwards by \code{mult}. If length 2, the lower limit
is expanded by \code{mult[1]} and the upper limit by \code{mult[2]}.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.
Giving a logical vector will separately control the expansion for the four
directions (top, left, bottom and right). The \code{expand} argument will be
recycled to length 4 if necessary. Alternatively, can be a named logical
vector to control a single direction, e.g. \code{expand = c(bottom = FALSE)}.}

\item{show_age_takeoff}{A logical value (default \code{TRUE}) to indicate
whether to display the ATGV line(s) on the plot.}

\item{show_age_peak}{A logical value (default \code{TRUE}) to indicate
whether to display the APGV line(s) on the plot.}

\item{show_age_cessation}{A logical value (default \code{TRUE}) to indicate
whether to display the ACGV line(s) on the plot.}

\item{show_vel_takeoff}{A logical value (default \code{FALSE}) to indicate
whether to display the TGV line(s) on the plot.}

\item{show_vel_peak}{A logical value (default \code{FALSE}) to indicate
whether to display the PGV line(s) on the plot.}

\item{show_vel_cessation}{A logical value (default \code{FALSE}) to indicate
whether to display the CGV line(s) on the plot.}

\item{each_object}{Optional logical (default \code{FALSE}) indicating whether
to return each plot object (combined and individual plots) or only the
combined when multiple plots are requested.}

\item{ncol}{Optional integer (default \code{NULL}) specifying the number of
columns for \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}. If \code{NULL}, then number of
columns is set as \code{2}.}

\item{print}{A logical value (default \code{TRUE}) to indicate whether
to print the plot object before returning it.}

\item{returndata}{A logical value (default \code{FALSE}) to indicate whether
to plot the data or return it as a \code{data.frame}.}

\item{returndata_add_parms}{A logical value (default \code{FALSE}) to specify
whether to add growth parameters to the returned \code{data.frame}. Ignored
when \code{returndata = FALSE}. Growth parameters are added when the
\code{opt} argument includes 'v' or 'V' and \code{apv = TRUE}.}

\item{parms_eval}{A logical value to specify whether or not to compute growth
parameters on the fly. This is for internal use only and is mainly needed
for compatibility across internal functions.}

\item{idata_method}{A character string specifying the interpolation method
(default \code{NULL}). The number of interpolation points is controlled by
the \code{ipts} argument.

Available options:
\itemize{
\item \code{'m1'}: Adapted from the \pkg{iapvbs} package (documented
\href{https://rdrr.io/github/Zhiqiangcao/iapvbs/src/R/exdata.R}{here}). This
method internally constructs the data frame based on the model
configuration. \emph{Note:} This method may fail if the model includes
covariates (especially in \code{univariate_by} models).
\item \code{'m2'}: Based on the \pkg{JMbayes} package (documented
\href{https://github.com/drizopoulos/JMbayes/blob/master/R/dynPred_lme.R}{here}).
This method uses the exact data frame from the model fit
(\code{fit$data}) and is generally more robust.
}

If \code{idata_method = NULL}, \code{'m2'} is automatically selected.
It is recommended to use \code{'m2'} if \code{'m1'} encounters errors
with covariate-dependent models.}

\item{parms_method}{A character string specifying the method used when
evaluating \code{parms_eval}. The default method is \code{getPeak}, which
uses the \code{\link[sitar:getPeak]{sitar::getPeak()}} function from the \code{sitar} package.
Alternatively, \code{findpeaks} uses the \code{findpeaks} function from the
\code{pracma} package. This parameter is for internal use and ensures
compatibility across internal functions.}

\item{verbose}{A logical argument (default \code{FALSE}) to specify whether
to print information collected during the setup of the object(s).}

\item{fullframe}{A logical value indicating whether to return a
\code{fullframe} object in which \code{newdata} is bound to the summary
estimates. Note that \code{fullframe} cannot be used with \code{summary =
FALSE}, and it is only applicable when \code{idata_method = 'm2'}. A
typical use case is when fitting a \code{univariate_by} model. This option
is mainly for internal use.}

\item{dummy_to_factor}{A named list (default \code{NULL}) to convert dummy
variables into a factor variable. The list must include the following
elements:
\itemize{
\item \code{factor.dummy}: A character vector of dummy variables to be
converted to factors.
\item \code{factor.name}: The name for the newly created factor variable
(default is \code{'factor.var'} if \code{NULL}).
\item \code{factor.level}: A vector specifying the factor levels.
If \code{NULL}, levels are taken from \code{factor.dummy}.
If \code{factor.level} is provided, its length must match
\code{factor.dummy}.
}}

\item{expose_function}{A logical value or \code{NULL} (default \code{FALSE}).
Controls whether Stan functions are exposed for post-processing.
\itemize{
\item \code{TRUE}: Explicitly exposes Stan functions saved from the model
fit. This is required when calculating \code{fit criteria} or
\code{bayes_R2} during model optimization.
\item \code{FALSE} (default): Stan functions are not exposed.
\item \code{NULL}: The setting is inherited from the original
\code{bsitar()} model fit configuration.
}
\strong{Note}: In the \code{\link[=optimize_model]{optimize_model()}} function, the default is
\code{NULL} (inheriting behavior), whereas other post-processing functions
default to \code{FALSE}.}

\item{usesavedfuns}{A logical value (default \code{NULL}) indicating whether
to use already exposed and saved Stan functions. This is typically set
automatically based on the \code{expose_functions} argument from the
\code{\link[=bsitar]{bsitar()}} call. Manual specification of \code{usesavedfuns} is rarely
needed and is intended for internal testing, as improper use can lead to
unreliable estimates.}

\item{clearenvfuns}{A logical value indicating whether to clear the exposed
Stan functions from the environment (\code{TRUE}) or not (\code{FALSE}). If
\code{NULL}, \code{clearenvfuns} is set based on the value of
\code{usesavedfuns}: \code{TRUE} if \code{usesavedfuns = TRUE}, or
\code{FALSE} if \code{usesavedfuns = FALSE}.}

\item{funlist}{A list (default \code{NULL}) specifying function names. This
is rarely needed, as required functions are typically retrieved
automatically. A use case for \code{funlist} is when \code{sigma_formula},
\code{sigma_formula_gr}, or \code{sigma_formula_gr_str} use an external
function (e.g., \code{poly(age)}). The \code{funlist} should include
function names defined in the \code{globalenv()}. For functions needing
both distance and velocity curves (e.g., \code{plot_curves(..., opt =
'dv')}), \code{funlist} must include two functions: one for the distance
curve and one for the velocity curve.}

\item{xvar}{A character string (default \code{NULL}) specifying the
\code{'x'} variable. Rarely used because \code{xvar} is inferred
internally. A use case is when conflicting variables exist (e.g.,
\code{sigma_formula}) and user wants to set a specific variable as
\code{'x'}.}

\item{difx}{A character string (default \code{NULL}) specifying the
\code{'x'} variable that should be used for manual differentiation of the
distance curve. Internally, the \code{xvar} is set as \code{difx} if
specified. The argument \code{difx} is evaluated only when \code{dpar =
'sigma'}, ignored otherwise. Note that argument \code{xvar} itself is rarely
used because \code{xvar} is inferred internally. A use case is when
conflicting variables exist (e.g., \code{sigma_formula}) and user wants to
set a specific variable as \code{'x'}.}

\item{idvar}{A character string (default \code{NULL}) specifying the
\code{'id'} variable. Rarely used because \code{idvar} is inferred
internally.}

\item{itransform}{A character string (default \code{NULL}) indicating the
variables names that are reverse transformed. Options are  \code{c("x",
"y", "sigma")}. The \code{itransform} is primarily used to get the
\code{xvar} variable at original scale i.e., \code{itransform = 'x'}. To
turn of all transformations, use \code{itransform = ""}. when
\code{itransform = NULL}, the appropriate transformation for \code{xvar} is
selected automatically. Note that when no match for \code{xvar} is found in
the \code{data,frame}, the \code{itransform} will be ignored within the
calling function, \code{'prepare_transformations()'}.}

\item{newdata_fixed}{An indicator specifying how to handle the format and
structure of the user-provided \code{newdata} before use. The available
options are:
\itemize{
\item \code{NULL} (default): Checks the data format and structure of
\code{newdata}, and applies \code{prepare_data2} and
\code{prepare_transformations} as needed.
\item \code{TRUE}: Returns \code{newdata} as-is, without any format
checking or transformation.
\item \code{0}: Same as \code{TRUE} — returns \code{newdata} without
checking the data format or applying \code{prepare_data2} and
\code{prepare_transformations}.
\item \code{1}: Checks the data format and, if needed, prepares it using
\code{prepare_data2}, but does not apply \code{prepare_transformations}.
\item \code{2}: Applies \code{prepare_transformations} only, assuming the
data format is already correct.
}
It is strongly recommended to leave both \code{newdata = NULL} and
\code{newdata_fixed = NULL} (the defaults). In this case, the data used
during model fitting is automatically retrieved, checked for the required
format, and transformed as needed via \code{prepare_data2} and
\code{prepare_transformations}. The flags \code{newdata_fixed = 0},
\code{1}, and \code{2} are intended primarily for internal use.}

\item{envir}{The environment used for function evaluation. The default is
\code{NULL}, which sets the environment to \code{parent.frame()}. Since
most post-processing functions rely on \pkg{brms}, it is recommended to set
\code{envir = globalenv()} or \code{envir = .GlobalEnv}, especially for
derivatives like velocity curves.}

\item{...}{Additional arguments passed to the \code{brms::fitted.brmsfit()}
and \code{brms::predict()} functions.}
}
\value{
A plot object (default) or a \code{data.frame} when \code{returndata
  = TRUE}.
}
\description{
The \strong{plot_curves()} function visualizes six different
types of growth curves using the \pkg{ggplot2} package. Additionally, it
allows users to create customized plots from the data returned as a
\code{data.frame}. For an alternative approach, the \code{\link[=get_predictions]{get_predictions()}}
function can be used, which not only estimates adjusted curves but also
enables comparison across groups using the \code{hypotheses} argument.
}
\details{
The \code{plot_curves()} function is a generic tool for visualizing up to six
types of growth curves:

\itemize{
\item Population-average distance curve (\code{opt = "d"})
\item Population-average velocity curve (\code{opt = "v"})
\item Individual-specific distance curves (\code{opt = "D"})
\item Individual-specific velocity curves (\code{opt = "V"})
\item Unadjusted individual growth curves (i.e., observed growth curves;
\code{opt = "u"})
\item Adjusted individual growth curves, adjusted for the model-estimated
random effects (\code{opt = "a"})
}

Users can request any combination of these plots by specifying the desired
curve types in the \code{opt} argument. For example, to request all six
plots, set \code{opt = "dvDVua"}. The plots are generated and assembled in
the same order as the characters appear in \code{opt}.

Internally, \strong{plot_curves()} calls the \code{\link[=growthparameters]{growthparameters()}} function
to estimate and summarize the distance and velocity curves, as well as to
compute growth parameters such as the age at peak growth velocity (APGV).
The function also calls \code{\link[=fitted_draws]{fitted_draws()}} or \code{\link[=predict_draws]{predict_draws()}} to make
inferences based on posterior draws. As a result, \strong{plot_curves()}
can plot either fitted or predicted curves. For more details, see
\code{\link[=fitted_draws]{fitted_draws()}} and \code{\link[=predict_draws]{predict_draws()}} to understand the difference between
fitted and predicted values.
}
\examples{
\donttest{
# Fit Bayesian SITAR model

# To avoid mode estimation which takes time, the Bayesian SITAR model is fit to
# the 'berkeley_exdata' and saved as an example fit ('berkeley_exfit').
# See 'bsitar' function for details on 'berkeley_exdata' and 'berkeley_exfit'.

# Check and confirm whether the model fit object 'berkeley_exfit' exists
berkeley_exfit <- getNsObject(berkeley_exfit)

model <- berkeley_exfit

# Population average distance and velocity curves with default options
plot_curves(model, opt = 'dv')

# Individual-specific distance and velocity curves with default options
# Note that \code{legendpos = 'none'} will suppress the legend positions.
# This suppression is useful when plotting individual-specific curves

plot_curves(model, opt = 'DV')

# Population average distance and velocity curves with APGV

plot_curves(model, opt = 'dv', apv = TRUE)

# Individual-specific distance and velocity curves with APGV

plot_curves(model, opt = 'DV', apv = TRUE)

# Population average distance curve, velocity curve, and APGV with CI bands
# To construct CI bands, growth parameters are first calculated for each
# posterior draw and then summarized across draws. Therefore,summary
# option must be set to FALSE

plot_curves(model, opt = 'dv', apv = TRUE, bands = 'dvp', summary = FALSE)

# Adjusted and unadjusted individual curves
# Note ipts = NULL (i.e., no interpolation of predictor (i.e., age) to plot a
# smooth curve). This is because it does not a make sense to interploate data
# when estimating adjusted curves. Also, layout = 'facet' (and not default
# layout = 'single') is used for the ease of visualizing the plotted
# adjusted and unadjusted individual curves. However, these lines can be
# superimposed on each other by setting the set layout = 'single'.
# For other plots shown above, layout can be set as 'single' or 'facet'

# Separate plots for adjusted and unadjusted curves (layout = 'facet')
plot_curves(model, opt = 'au', ipts = NULL, layout = 'facet')

# Superimposed adjusted and unadjusted curves (layout = 'single')
plot_curves(model, opt = 'au', ipts = NULL, layout = 'single')

}

}
\seealso{
\code{\link[=growthparameters]{growthparameters()}} \code{\link[=fitted_draws]{fitted_draws()}} \code{\link[=predict_draws]{predict_draws()}}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
