% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check_.bsitar.R
\name{pp_check_.bsitar}
\alias{pp_check_.bsitar}
\alias{pp_check_}
\title{Posterior predictive distribution check for \code{bsitar} object}
\usage{
\method{pp_check_}{bsitar}(model, resp = NULL, envir = parent.frame(), ...)

pp_check_(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar} function for details).}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}
function. Please see \code{brms::pp_check.brmsfit} for details on
various options available.}
}
\value{
A ggplot object that can be further customized using the
ggplot2 package.
}
\description{
Perform posterior predictive checks with the help
of the \pkg{bayesplot} package.
}
\details{
The \code{pp_check_} function is a wrapper around
the \link[brms:loo.brmsfit]{brms::loo} and works exactly the same way as
\link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}.
}
\examples{

\dontrun{
pp_check_(model)
}
}
