% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check_bgm.R
\name{pp_check_bgm.bgmfit}
\alias{pp_check_bgm.bgmfit}
\alias{pp_check_bgm}
\title{Posterior predictive distribution check for \code{bgmfit} object}
\usage{
\method{pp_check_bgm}{bgmfit}(model, resp = NULL, envir = globalenv(), ...)

pp_check_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bgmfit} function for details).}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}
function. Please see \code{brms::pp_check.brmsfit} for details on
various options available.}
}
\value{
A ggplot object that can be further customized using the
ggplot2 package.
}
\description{
Perform posterior predictive checks with the help
of the \pkg{bayesplot} package.
}
\details{
The \code{pp_check_bgm} function is a wrapper around
the \link[brms:loo.brmsfit]{brms::loo} and works exactly the same way as
\link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}.
}
\examples{

# Fit Bayesian SITAR model 
# To avoid running the model which takes some time, model fit to the
# \code{berkeley_mdata} has already been saved as berkeley_mfit.rda object.
# Please see \code{bgm} examples.

model <- berkeley_mfit

pp_check_bgm(model)

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
