% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Set up data for bsitar model}
\usage{
prepare_data(
  data,
  response_variable,
  uvarby = NA,
  mvar = FALSE,
  xfuns = NULL,
  yfuns = NULL
)
}
\arguments{
\item{data}{An input data frame}

\item{response_variable}{The name(s) of the outcome(s) variables. Must be a
single name except when fitting a multivariate model.}

\item{uvarby}{An optional (default \code{NA}) to specify the indicator
variable for fitting univariate-by-subgroup model. See \code{univariate_by}
argument in the \code{\link[=bsitar]{bsitar()}} function. If not \code{NA}, then it
should be a valid factor variable present in the \code{data}.}

\item{mvar}{A logical (default \code{FALSE}) to specify the the multivariate
model. See \code{multivariate} argument in the \code{\link[=bsitar]{bsitar()}} function.}

\item{xfuns}{Optional name(s) of the transformation function(s) applied to
the predictor variable (typically age). Default \code{NULL}.}

\item{yfuns}{Optional name(s) of the transformation function(s) applied to
the outcome variable. Default \code{NULL}.}
}
\value{
A data frame with necessary information added a attributes.
}
\description{
Set up data for bsitar model
}
\examples{
\dontrun{
data(heights)
prepare_data(data = heights, response_variable = heights)
}
}
