% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_formula.R
\name{prepare_formula}
\alias{prepare_formula}
\title{Prepare formula for fitting Bayesian SITAR growth curve model}
\usage{
prepare_formula(x, y, id, knots, nknots, data, internal_formula_args)
}
\arguments{
\item{x}{vector of predictor (typically age in years).}

\item{y}{vector of outcome (i.e., repeated growth measurements).}

\item{id}{a factor variable identifying the groups (typically individuals).}

\item{knots}{vector of values for knots.}

\item{nknots}{an integer specifying the number of knots.}

\item{data}{data frame containing variables \code{x}, \code{y} and \code{id}.}

\item{internal_formula_args}{Other internal arguments passed from the
\link{bsitar} to the \code{prepare_formula}).}
}
\value{
An object of class \code{brmsformula}, which is a \code{list}
containing formulas
}
\description{
The \code{prepare_formula}) prepares \code{brms::brmsformual} which is
passed on to the \link{bsitar} function. For univariate-by-
subgroup model (specified by using the \code{univariate_by}) and
multivariate model (specified by using the \code{multivariate}),
the \code{x}, \code{y}, \code{id}, \code{knots}, \code{nknots}, are
automatically set to match the sub-model(s). See \code{brms::brmsformual}
for details.
}
\seealso{
\link[brms:brm]{brms::brm} \link[brms:brmsformula]{brms::brmsformula}
}
