% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_function.R
\name{prepare_function}
\alias{prepare_function}
\title{Prepare Stan function for fitting Bayesian SITAR growth curve model}
\usage{
prepare_function(x, y, id, knots, nknots, data, internal_function_args)
}
\arguments{
\item{x}{vector of predictor (typically age in years).}

\item{y}{vector of outcome (i.e., repeated growth measurements).}

\item{id}{a factor variable identifying the groups (typically individuals).}

\item{knots}{vector of values for knots.}

\item{nknots}{an integer specifying the number of knots.}

\item{data}{data frame containing variables \code{x}, \code{y} and \code{id}.}

\item{internal_function_args}{Other internal arguments passed from the
\link{bsitar} to the \code{prepare_formula}).}
}
\value{
An character string which later evaluated to a custom function
and inserted into the Stan's functions block.
}
\description{
The \code{prepare_function}) prepares custom Stan function  which is passed
on to the \link{bsitar} function. For univariate-by-
subgroup model (specified by using the \code{univariate_by}) and
multivariate model (specified by using the \code{multivariate}),
the \code{x}, \code{y}, \code{id}, \code{knots}, \code{nknots}, are
automatically matched with the sub-model(s).
}
