% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_initials.R
\name{prepare_initials}
\alias{prepare_initials}
\title{Prepare initials for fitting Bayesian SITAR growth curve model}
\usage{
prepare_initials(init_argument, init_internal_args)
}
\arguments{
\item{init_argument}{A list containing the initial arguments specified in
the \code{\link[=bgm]{bgm()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function via \code{\link[=prepare_priors]{prepare_priors()}} to
the \code{prepare_initials}. See \code{\link[=bgm]{bgm()}} function for details on
initials arguments.}

\item{init_internal_args}{An internal argument list created and passed from
the \code{\link[=prepare_priors]{prepare_priors()}} function to this \code{prepare_initials}
function.}
}
\value{
A list of initial values.
}
\description{
For univariate-by-subgroup (specified by using the \code{univariate_by})
and multivariate (specified by using the \code{multivariate}) models, each
argument is automatically matched with the sub-model(s). See
\code{\link[=bgm]{bgm()}} for details on \code{univariate_by} and
\code{multivariate}) arguments.
}
