% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_priors.R
\name{prepare_priors}
\alias{prepare_priors}
\title{Prepare priors for fitting Bayesian SITAR growth curve model}
\usage{
prepare_priors(
  prior_argument,
  prior_data,
  prior_data_internal,
  prior_internal_args,
  init_arguments,
  init_data,
  init_data_internal,
  init_args_internal
)
}
\arguments{
\item{prior_argument}{A list containing the prior arguments specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.}

\item{prior_data}{An optional argument (as named list) specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.
The \code{prior_data} used to pass value for priors.
See \code{\link[=bsitar]{bsitar()}} function, \code{prior_data} for details.}

\item{prior_data_internal}{An internal argument (as named list) specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.}

\item{prior_internal_args}{An internal argument list that is passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{set_priors_initials}
and is used in setting the priors.}

\item{init_arguments}{A list containing the initial arguments specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.}

\item{init_data}{An optional argument (as named list) specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.
The \code{init_data} is used for setting the initials.}

\item{init_data_internal}{An internal argument (as named list) specified in
the \code{\link[=bsitar]{bsitar()}} function and then passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{prepare_priors}.}

\item{init_args_internal}{An internal argument list that is passed from the
\code{\link[=set_priors_initials]{set_priors_initials()}} function to the \code{set_priors_initials}
and is used in setting the initials.}
}
\value{
An object of class \code{brmsprior}. See \code{brmsprior} function
for more details.
}
\description{
For univariate-by-subgroup (specified by using the \code{univariate_by})
and multivariate (specified by using the \code{multivariate}) models, each
argument is automatically matched with the sub-model(s). See
\code{\link[=bsitar]{bsitar()}} for details on \code{univariate_by} and
\code{multivariate}) arguments.
}
\seealso{
\link[brms:brm]{brms::brm} \link[brms:set_prior]{brms::prior}
}
