% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors_initials.R
\name{set_priors_initials}
\alias{set_priors_initials}
\title{Set priors and initials for fitting Bayesian SITAR growth curve model}
\usage{
set_priors_initials(
  a_prior_beta,
  b_prior_beta,
  c_prior_beta,
  d_prior_beta,
  e_prior_beta,
  f_prior_beta,
  s_prior_beta,
  a_cov_prior_beta,
  b_cov_prior_beta,
  c_cov_prior_beta,
  d_cov_prior_beta,
  e_cov_prior_beta,
  f_cov_prior_beta,
  s_cov_prior_beta,
  a_prior_sd,
  b_prior_sd,
  c_prior_sd,
  d_prior_sd,
  e_prior_sd,
  f_prior_sd,
  a_cov_prior_sd,
  b_cov_prior_sd,
  c_cov_prior_sd,
  d_cov_prior_sd,
  e_cov_prior_sd,
  f_cov_prior_sd,
  gr_prior_cor,
  sigma_prior_cor,
  sigma_prior_beta,
  sigma_cov_prior_beta,
  sigma_prior_sd,
  sigma_cov_prior_sd,
  rsd_prior_sigma,
  dpar_prior_sigma,
  dpar_cov_prior_sigma,
  autocor_prior_acor,
  autocor_prior_unstr_acor,
  mvr_prior_rescor,
  prior_data = NULL,
  prior_data_internal = NULL,
  prior_args_internal = NULL,
  init_arguments = NULL,
  init_data = NULL,
  init_data_internal = NULL,
  init_args_internal = NULL,
  custom_order_prior_str = NULL
)
}
\arguments{
\item{a_prior_beta}{Specify priors for the fixed effect parameter, \code{a}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{b_prior_beta}{Specify priors for the fixed effect parameter, \code{b}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{c_prior_beta}{Specify priors for the fixed effect parameter, \code{c}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{d_prior_beta}{Specify priors for the fixed effect parameter, \code{d}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{e_prior_beta}{Specify priors for the fixed effect parameter, \code{e}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{f_prior_beta}{Specify priors for the fixed effect parameter, \code{f}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{s_prior_beta}{Specify priors for the fixed effect parameter, \code{s}.
See \code{\link[=bsitar]{bsitar()}} for details.}

\item{a_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{a}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{b_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{b}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{c_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{c}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{d_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{d}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{e_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{e}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{f_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{f}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{s_cov_prior_beta}{Specify priors for the covariate(s) included for the
fixed effect parameter, \code{s}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{a_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{a}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{b_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{b}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{c_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{c}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{d_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{d}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{e_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{e}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{f_prior_sd}{Specify prior on the standard deviation (sd) of random
effect parameter, \code{f}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{a_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{a}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{b_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{b}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{c_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{c}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{d_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{d}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{e_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{w}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{f_cov_prior_sd}{Specify prior on the standard deviation (sd) for the
covariate(s) included in the random effect parameter, \code{f}. See
\code{\link[=bsitar]{bsitar()}} for details.}

\item{gr_prior_cor}{Specify prior on the correlation for group level random
effect parameters. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{sigma_prior_cor}{Specify prior on the correlation for distribution level
random effect parameters. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{sigma_prior_beta}{Specify prior on the the distributional fixed effect
parameter, \code{sigma}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{sigma_cov_prior_beta}{Specify prior on the covariate(s) included in the
distributional fixed effect parameter, \code{sigma}.  See \code{\link[=bsitar]{bsitar()}}
for details.}

\item{sigma_prior_sd}{Specify prior on the standard deviation (sd) of
distributional random effect parameter, \code{sigma}. See \code{\link[=bsitar]{bsitar()}}
for details.}

\item{sigma_cov_prior_sd}{Specify prior on the standard deviation (sd) of
covariate(s) included in the distributional random effect parameter,
\code{sigma}. See \code{\link[=bsitar]{bsitar()}} for details.}

\item{rsd_prior_sigma}{Specify prior on the residual standared deviation
parameter, \code{sigma}, See \code{\link[=bsitar]{bsitar()}} for details,}

\item{dpar_prior_sigma}{Specify prior on the distributional parameter,
\code{sigma} (which is same as residual standared deviation for Gaussian
distribution). See \code{\link[=bsitar]{bsitar()}} for details,}

\item{dpar_cov_prior_sigma}{Specify prior for the covariate(s) included in the
distributional parameter, \code{sigma} (which is same as residual standard
deviation for Gaussian distribution).See \code{\link[=bsitar]{bsitar()}} for details,}

\item{autocor_prior_acor}{Specify priors on the the autocorrelation parameters
\code{ar}, \code{ma} and \code{arma}. See \code{\link[=bsitar]{bsitar()}} for details,}

\item{autocor_prior_unstr_acor}{Specify priors on the the unstructured
autocorrelation parameter. See \code{\link[=bsitar]{bsitar()}} for details,}

\item{mvr_prior_rescor}{Specify priors on the the residual correlation
parameter for multivariate model. See \code{\link[=bsitar]{bsitar()}} for details,}

\item{prior_data}{An optional argument (as named list) that can pass value for
prior. See \code{\link[=bsitar]{bsitar()}} for details,}

\item{prior_data_internal}{An internal data frame (named list) used to pass on
the relevant information on priors from the \code{\link[=bsitar]{bsitar()}} function to
the \code{set_priors_initials}.}

\item{prior_args_internal}{An internal argument list that is passed from the
\code{\link[=bsitar]{bsitar()}} function to the \code{set_priors_initials} and is used
for setting the priors.}

\item{init_arguments}{A list containing all the init arguments specified in
the \code{\link[=bsitar]{bsitar()}} function and now passed on to the
\code{set_priors_initials}.}

\item{init_data}{An optional data argument (named list) used to pass initial
values. See \code{\link[=bsitar]{bsitar()}} function, \code{prior_data} for details.}

\item{init_data_internal}{An internal data frame (named list) to pass on the
relevant information on initials from the \code{\link[=bsitar]{bsitar()}} function to the
\code{set_priors_initials}.}

\item{init_args_internal}{An internal argument list that is passed from the
\code{\link[=bsitar]{bsitar()}} function to the \code{set_priors_initials} and is used
for setting the initials.}

\item{custom_order_prior_str}{An internal argument that is passed from the
\code{\link[=bsitar]{bsitar()}} function to the \code{set_priors_initials} when setting
the priors for the model with hierarchy level 3 and beyond. See
\code{\link[=bsitar]{bsitar()}} for details,}
}
\value{
An object of class \code{brmsprior} (See \code{brmsprior}). In
addition to the priors, the returned object contains a list of initial
values.
}
\description{
The \code{set_priors_initials}) sets priors and initials values
which are passed from the \link{bsitar} function to
\code{set_priors_initials}. For univariate-by- subgroup model (specified by
using the \code{univariate_by}) and multivariate model (specified by using
the \code{multivariate}), each argument is automatically matched with the
sub-model(s).
}
