% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R, R/update.R
\name{optimize}
\alias{optimize}
\alias{update.bsitar}
\alias{update}
\title{Update \pkg{bsitar} models}
\usage{
optimize(model, ...)

\method{update}{bsitar}(model, newdata = NULL, recompile = T, ...)

update(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{...}{Other arguments passed to \code{\link{brms}}.}

\item{newdata}{newdata Optional \code{data.frame} to update the model with
new data. Note that data-dependent default priors will not be updated
automatically.}

\item{recompile}{A logical to indicate whether the Stan model should be
recompiled. When \code{NULL} (the default), \code{update} tries to figure
out internally, if recompilation is necessary. Setting it to \code{FALSE}
will cause all Stan code changing arguments to be ignored.}
}
\value{
An updated object of class \code{brmsfit, bsiatr}, that contains the
posterior draws and other useful information about the model.
}
\description{
Update \pkg{bsitar} models
}
\examples{
\dontrun{
data(heights)
data_males <- heights \%>\% filter(sex == 'Male')
fit_males <- bsitar(x=age, y=height, id=id, data=heights, df=4)
fit_males2 <- update(df=5)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
