% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bsitar.R
\name{update_bsitar.bsitar}
\alias{update_bsitar.bsitar}
\alias{update_bsitar}
\title{Update \pkg{bsitar} model}
\usage{
\method{update_bsitar}{bsitar}(model, data = NULL, recompile = NULL, ...)

update_bsitar(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{data}{An optional \code{data.frame} to be used when updating the model.
If \code{NULL} (default), the same same data used for original model fit is
used. Note that data-dependent default priors will not be updated
automatically.}

\item{recompile}{A logical to indicate whether the Stan model should be
recompiled. When \code{NULL} (the default), \code{update_bsitar} tries to figure
out internally, if recompilation is necessary. Setting it to \code{FALSE}
will cause all Stan code changing arguments to be ignored.}

\item{...}{Other arguments passed to \code{\link{brms}}.}
}
\value{
An updated object of class \code{brmsfit, bsiatr}, that contains the
posterior draws and other useful information about the model.
}
\description{
Update \pkg{bsitar} model
}
\examples{
\dontrun{
data(heights)
data_males <- heights \%>\% filter(sex == 'Male')
fit_males <- bsitar(x=age, y=height, id=id, data=heights, df=4)
fit_males2 <- update_bsitar(df=5)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
