% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_exfit.R
\docType{data}
\name{berkeley_exfit}
\alias{berkeley_exfit}
\title{Model fit to the Berkeley Child Guidance Study Data for females}
\format{
Model fit comprising summary of posterior draws.
}
\usage{
berkeley_exfit
}
\value{
An object of class \code{bgmfit} with posterior draws.
}
\description{
Bayesian SITAR model fit to the \link{berkeley_exdata}
data (70 females, 8 to 18 years of age).
}
\details{
Data details are provided in the \link{berkeley_exdata}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
