% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_mdata.R
\docType{data}
\name{berkeley_mdata}
\alias{berkeley_mdata}
\title{Berkeley Child Guidance Study Data for males}
\format{
A data frame with 902 observations on the following 3 variables:
\describe{
\item{id}{factor variable}
\item{age}{years, numeric vector}
\item{height}{cm, numeric vector}
}
}
\usage{
berkeley_mdata
}
\value{
A data frame with 3 columns.
}
\description{
A subset of the \link{berkeley} data that contains
longitudinal growth data for 20 randomly selected males (6 to 20
years of age).
}
\details{
A detailed description of the full data is provided in the
\link{berkeley} data.
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
