% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_mdata.R
\docType{data}
\name{berkeley_mdata}
\alias{berkeley_mdata}
\title{Berkeley Child Guidance Study Data for males}
\format{
A data frame with 902 observations on the following 10 variables:
\describe{
\item{id}{factor with levels 201-278}
\item{age}{years, numeric vector}
\item{height}{cm, numeric vector}
}
}
\usage{
berkeley_mdata
}
\value{
A data frame with 10 columns.
}
\description{
The Berkeley Child Guidance Study data for males is subset of
the Berkeley Child Guidance Study data \code{berkeley} that contains
longitudinal growth data for 136 children (66 males and 70 females) from
birth to 21 years. The subset data \code{berkeley_mdata} includes growth
measurements on males between 6 and 20 years of age.
}
\details{
The data were originaly provided as an appendix to the book by
\insertCite{Tuddenham1954;textual}{bsitar}. The same data has been used in
the \emph{sitar} \insertCite{R-sitar}{bsitar} and the \emph{fda}
\insertCite{R-fda}{bsitar} packages after correcting the transcription
errors. The data included in the \strong{bsitar} package is the same as
\code{berkeley} data used in the \emph{sitar} package.

A detailed description of the data including the frequency of measurements
per year is provided in the the \emph{sitar} package.
\insertCite{R-sitar}{bsitar} and the \link{berkeley_mdata} included in
this \code{bsitar} package.
}
\references{
\insertAllCited{}
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
