% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkeley_mfit.R
\docType{data}
\name{berkeley_mfit}
\alias{berkeley_mfit}
\title{Model fit to the Berkeley data for males}
\format{
Model fit comprising summary of posterior draws.
}
\usage{
berkeley_mfit
}
\value{
An object of class \code{bgmfit} with posterior draws.
}
\description{
Bayesian SITAR model fit to the \link{berkeley_mdata}
data (20 males, 6 to 20 years of age).
}
\details{
Data details are provided in the \link{berkeley_mdata}.
}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
\keyword{datasets}
