% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_functions_bgm.R
\name{expose_functions_bgm}
\alias{expose_functions_bgm}
\title{Expose Stan function for post-processing of posterior sample obtained
from the model fit}
\usage{
expose_functions_bgm(model, scode = NULL, expose = FALSE, select_model = NULL)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{scode}{An option argument specifying the code with the user-defined
Stan function.}

\item{expose}{A logical (default \code{TRUE}) to indicate whether to expose
functions and add to the global environment.}

\item{select_model}{A string (default \code{NULL}) to indicate the model.}
}
\value{
An object of class \code{bgmfit} with exposed
user-defined Stan functions (when \code{expose=TRUE}).
}
\description{
Expose Stan function for post-processing of posterior sample obtained
from the model fit
}
\examples{

# The examples below show the use of *conditional_effects_bgm* to plot  
# the population average and individual-specific distance and velocity 
# curves.

# Fit Bayesian SITAR model 
# To avoid running the model which takes some time, model fit to the
# \code{berkeley_mdata} has already been saved as berkeley_mfit.rda object.
# Please see \code{bgm} examples.

model <- berkeley_mfit

\donttest{
expose_functions_bgm(model, expose = TRUE)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
