% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get_object.R
\name{getNsObject}
\alias{getNsObject}
\title{Check and get namespace object if exists}
\usage{
getNsObject(object, namespace = NULL, envir = NULL)
}
\arguments{
\item{object}{An object to be retrieved. Note that \code{object} must be a
symbol and not a character string.}

\item{namespace}{A character string specifying the namespace to be checked.}

\item{envir}{An environment to be used (default \code{NULL}).}
}
\value{
An object of same class as input \code{object}.
}
\description{
Check and get namespace object if exists
}
\examples{

\donttest{
# Check whether model fit object 'berkeley_exfit' exists
 berkeley_exfit <- getNsObject(berkeley_exfit)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
