% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gparameters.R
\name{gparameters.bsitar}
\alias{gparameters.bsitar}
\alias{gparameters}
\title{Growth parameter estimation for \code{bsitar} model}
\usage{
\method{gparameters}{bsitar}(
  model,
  resp = NULL,
  ndraws = NULL,
  draw_ids = NULL,
  newdata = NULL,
  summary = TRUE,
  robust = FALSE,
  re_formula = NA,
  peak = TRUE,
  takeoff = FALSE,
  trough = FALSE,
  estimation_method = "fitted",
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty",
  incl_autocor = TRUE,
  numeric_cov_at = NULL,
  aux_variables = NULL,
  levels_id = NULL,
  avg_reffects = NULL,
  conf = 0.95,
  ipts = NULL,
  xrange = NULL,
  seed = 123,
  future = FALSE,
  future_session = "multisession",
  cores = NULL,
  ...
)

gparameters(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{resp}{Optional names of response variables. If specified, predictions
are performed only for the specified response variables.}

\item{ndraws}{Positive integer indicating the number of posterior draws to be
used in estimation. If \code{NULL} (default), all draws are used.}

\item{draw_ids}{An integer vector specifying the posterior draws
o be used. If NULL (the default), all draws are used.}

\item{newdata}{An optional data.frame for which to evaluate predictions. If
\code{NULL} (default), the original data of the model is used.}

\item{summary}{A logical (default \code{TRUE}) to indicate whether only the
Estimate should be returned or Estimate along with SE and CI should be
computed. Setting this option to \code{FALSE} will reduce the computation
time but no SE or CI estimates will be available.}

\item{robust}{If \code{FALSE} (the default) the mean is used as the measure
of central tendency and the standard deviation as the measure of
variability. If \code{TRUE}, the median and the median absolute deviation
(MAD) are applied instead. Ignored if summary is \code{FALSE}}

\item{re_formula}{Option to indicate whether or not to include the
individual/group-level effects in the estimation. When \code{NA} (default),
the individual-level effects are excluded and therefore population average
growth parameters are computed. When \code{NULL}, individual-level effects
are included in the computation and hence the growth parameters estimates
returned are individual-specific. In both sitations, (i.e,, \code{NA} or
\code{NULL}), continuous and factor covariate(s) are appropriately
included. When the continuous covariates by default are set to their means
(see \code{numeric_cov_at} for details), the population average as well as
individual-specific growth parameter estimates are returned for each level
of the factor cocariate.}

\item{peak}{Optional logical specifying whether or not to calculate the age
at peak velocity from the velocity curve. If \code{TRUE} (default), age at
peak velocity (APGV) and the peak velocity (PGV) are calculated along with
the uncertainty (SE and CI) around the APGV and PGV parameters. See
@details for further information.}

\item{takeoff}{Optional logical (default \code{FALSE}) specifying whether or
not to calculate the age at takeoff velocity from the velocity curve. If
\code{TRUE}, age at takeoff velocity (ATGV) and the takeoff growth velocity
(TGV) are  calculated along with the the uncertainty (SE and CI) around the
ATGV and TGV parameters. See @details for further information.}

\item{trough}{Optional logical (default \code{FALSE}) specifying whether or
not to calculate the age at takeoff velocity from the velocity curve. If
\code{TRUE}, age at cessation of growth velocity (ACGV) and the cessation
growth velocity (CGV) are  calculated along with the the uncertainty (SE
and CI) around the ACGV and CGV parameters. See @details for further
information.}

\item{estimation_method}{A character string to specify the estimation method
used to calculate velocity from the posterior draws. The \code{'fitted'}
method internally calls the \link{fitted_.bsitar} function whereas the
option \code{'predict'} calls the \link{predict_.bsitar} function. See
\link[brms:fitted.brmsfit]{brms::fitted.brmsfit} and \link[brms:predict.brmsfit]{brms::predict.brmsfit} for derails on fitted
versus predict estimations as they are the backend functions called by the
\link{fitted_.bsitar} and \link{predict_.bsitar} functions.}

\item{allow_new_levels}{A flag indicating if new levels of group-level
effects are allowed (defaults to \code{FALSE}). Only relevant if
\code{newdata} is provided.}

\item{sample_new_levels}{Indicates how to sample new levels for grouping
factors specified in \code{re_formula}. This argument is only relevant if
\code{newdata} is provided and \code{allow_new_levels} is set to
\code{TRUE}. If \code{"uncertainty"} (default), each posterior sample for a
new level is drawn from the posterior draws of a randomly chosen existing
level. Each posterior sample for a new level may be drawn from a different
existing level such that the resulting set of new posterior draws
represents the variation across existing levels. If \code{"gaussian"},
sample new levels from the (multivariate) normal distribution implied by the
group-level standard deviations and correlations. This options may be useful
for conducting Bayesian power analysis or predicting new levels in
situations where relatively few levels where observed in the old_data. If
\code{"old_levels"}, directly sample new levels from the existing levels,
where a new level is assigned all of the posterior draws of the same
(randomly chosen) existing level.}

\item{incl_autocor}{A flag indicating if correlation structures originally
specified via \code{autocor} should be included in the predictions.
Defaults to \code{TRUE}.}

\item{numeric_cov_at}{An optional argument to specify the value of continuous
covariate(s) before calling the \link{fitted_.bsitar} and
\link{predict_.bsitar} functions. The default \code{NULL} option set the
continuous covariate(s) at mean. Alternatively, a named list can be
supplied to manualy set the values. For example, \code{numeric_cov_at =
  list(xx = 2)} will set the continuous covariate varibale 'xx' at 2. The
argument \code{numeric_cov_at} is ignored when no continuous covariate is
included in the model.}

\item{aux_variables}{An optional argument to specify the variables to be
passed to the \code{ipts} argument. This is useful when fitting location
scale models and the measurement error models.}

\item{levels_id}{An optional argument to specify the ids for hierarchical
model (default \code{NULL}. It is used only when a with 3 or higher level
structures are fit. For two level model, the the id for second level is
automatically inferred from the fitted model. Even for 3 or higher level
model, ids are inferred from the fitted model but under the assumption that
hierarchy is specified from lower to upper levels i.e, id, study assuming
that id is nested within the studies. However, it is not gaunated that
these ids are sorted correctly. Therefore, it is better to set it manually.}

\item{avg_reffects}{An optional argument (default \code{NULL} to calculate
(marginal/average) distance and velocity curves (as well as APGV and PGV).
If specified, it must be set as a named list indicating the \code{over} and
the \code{by} arguments e.g., \code{avg_reffects = list(by = 'study', over
  = 'age'}. The \code{over} is typically age and is used to average over the
random effects. The second argument is \code{by} that specifies the factor
variable by which \code{over} is executed.}

\item{conf}{A numeric value (default \code{0.95}) to compute CI. Internally,
this is translated into a paired probability values as \code{c((1 - conf) /
  2, 1 - (1 - conf) / 2)}. For \code{conf = 0.95}, this will compute 95\% CI
with CI varibales named as Q.2.5 and Q.97.5.}

\item{ipts}{A numeric value to interpolate the predictor value to get a
smooth velocity curve. The \code{NULL} (default) will return original
values whereas a positive real value (e.g., \code{ipts = 10}) will
interpolate the predictor. It is important to note that these
interpolations do not alter the range of predictor when calculating
population average and the individual specific velocity curves.}

\item{xrange}{An indicator to set the predictor range (i.e., age) when
executing the interpolation via \code{ipts}. The default \code{NULL} sets
the individual specific predictor range whereas code
\code{xrange = 1} sets same range for all individuals within the higher order grouping variable (e.g., study). Code \code{xrange
  = 2} sets the identical range across the entire sample. Lastly, a paired numeric
values can be supplied e.g., \code{xrange = c(6, 20)} will set the range
between 6 and 20.}

\item{seed}{An integer (default \code{123}) that is passed to the estimation
method.}

\item{future}{A logical (default \code{FALSE}) to specify whether or not to
perform parallel computations. If set to \code{TRUE}, the
\link[future.apply:future_lapply]{future.apply::future_sapply} function is used for summarizing the draws.}

\item{future_session}{A character string to set the session type when
\code{future = TRUE}. The \code{'multisession'} (default) options sets the
multisession whereas the \code{'multicore'} set the multicore session. Note
that multicore session are not supported on Windows systems. For more
details, see \link{future.apply} and \link{future}.}

\item{cores}{Number of cores to be used when running the parallel
computations by setting the option \code{future = TRUE}. On non-Windows,
systems this argument can be set globally via the mc.cores option. For the
default \code{NULL} option, the number of cores are set automatically by
calling the \code{\link[future:re-exports]{future::availableCores()}}. The number of cores used are the
maximum number of cores avaialble minus one, i.e.,
\code{future::availableCores() - 1}.}

\item{...}{Further arguments passed to \code{brms::fitted} and
\code{brms::predict} functions that control several aspects of data
validation and prediction. See \link[brms:fitted.brmsfit]{brms::fitted.brmsfit} and
\link[brms:predict.brmsfit]{brms::predict.brmsfit} for details.}
}
\value{
a data frame with five columns when \code{summary = TRUE}, and two
columns when \code{summary = False} (assuming \code{re_formual = NULL}).
The first two columns common to both these approaches \code{summary =
  TRUE/False} are 'Parameter' and 'Estimate' which indicates the name of the
growth parameter (e.g., APGV, PGV etc), and its value. When \code{summary =
  TRUE}, three additional columns are added, the 'Est.Error' and a pair of
columns showing the CI. The CI columns are named as Q with appropriate
suffix indicating the percentiles used to construct these intervals (such
as Q.2.5 and Q.97.5 where 2.5 and 97.5 are the 0.025 and 0.975 percentiles
used to compute by the 95\% CI by calling the quantile function. When
\code{re_formual = NULL}, an additional column is added that shows the
individuals/groups corresponding to the estimates.
}
\description{
The \code{gparameters} computes the growth parameter estimates
and the uncertainty (standard errors, SE and credible intervals, CI) for
population average and individual-specific parameters such as the peak
growth velocity and the age at peak growth velocity.
}
\details{
The \code{gparameters} functions first call the appropriate function
(fitted or predict) to estimate velocity curve for each draw from the
posterior distribution and then estimates various growth parameters such as
age at peak growth velocity (APGV), peak growth velocity (PGV), age at
takeoff growth velocity (ATGV), takeoff growth velocity (TGV), age at
cessation of growth velocity (ACGV), and the cessation growth velocity
(CGV). The growth parameters APGV and PGV are estimated by calling the
\link[sitar:getPeakTrough]{sitar::getPeak} function whereas the ATGV and TGV are estimated by using
the \link[sitar:getPeakTrough]{sitar::getTakeoff} function. The \link[sitar:getPeakTrough]{sitar::getTrough} function is used
to estimates ACGV and CGV parameters. These parameters obtained for each
posterior draw are then summarized appropriately to get the estimate and
the uncertainty (SE and CI) around these estimates. Please note that it is
not always possible to estimate cessation and takeoff growth parameters
when there are no distinct pre-peak or post-peak troughs.
}
\examples{
\dontrun{
# Population average APGV and PGV
gparameters(model, re_formula = NA)

#' # Population average APGV, PGV, ATGV, TGV
gparameters(model, re_formula = NA, peak = TRUE, takeoff = TRUE)

# Individual-specific APGV, PGV, ATGV, TGV
gparameters(model, re_formula = NULL, peak = TRUE, takeoff = TRUE)

}

}
