% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights_fit.R
\docType{data}
\name{heights_fit}
\alias{heights_fit}
\title{Model fit to serial measurements of height on 12 girls}
\format{
Data frame analysed have the following 3 variables:
\describe{
\item{id}{factor of subject ids (levels 1:12)}
\item{age}{vector of ages (years)}
\item{height}{vector of heights (cm)}
}
}
\usage{
heights_fit
}
\value{
A object of class \code{bgmfit} with posterior draws.
}
\description{
Model is fit to data comprise of height measurements of 12
girls from the Chard Growth Study measured twice a year between 8 and
16 years of age.
}
\details{
Data details are provided in the the \emph{sitar} package
\insertCite{R-sitar}{bsitar}. The data included in the \strong{bsitar} package is
the same as \code{heights} data used in the \emph{sitar} package.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
