% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper-1.R
\name{is.bgmfit}
\alias{is.bgmfit}
\title{Checks if argument is a \code{bgmfit} object}
\usage{
is.bgmfit(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{bgmfit} object
}
