% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_bsitar.R
\name{optimize_bsitar.bsitar}
\alias{optimize_bsitar.bsitar}
\alias{optimize_bsitar}
\title{Optimize \pkg{bsitar} model}
\usage{
\method{optimize_bsitar}{bsitar}(
  model,
  data = NULL,
  optimize = list(x = c("NULL", "log", "sqrt"), y = c("NULL", "log", "sqrt")),
  exclude_default = FALSE,
  add_fit_criteria = NULL,
  add_fit_bayes_R = NULL,
  ...
)

optimize_bsitar(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.}

\item{data}{An optional \code{data.frame} to be used when optimising the
model. If \code{NULL} (default), the same same data used for original model
fit is used. Note that data-dependent default priors will not be updated
automatically.}

\item{optimize}{A list specifying the transformations  of predictor
(typically \code{age}) variable (via \code{xvar}) and the outcome (via
\code{yvar}). The options for both \code{xvar} and \code{yvar}) are 'NULL',
'log' and  'sqrt' or their combinations. The default
\code{optimize = list(x = c('NULL', 'log',  'sqrt'),
   y = c('NULL', 'log',  'sqrt'))} is to explore all possible combinations of
'NULL', 'log' and  'sqrt'.}

\item{exclude_default}{A logical (deafult \code{FALSE}) to indicate whether
transformations (\code{xvar} and \code{yvar}) used in the original model
fit should be excluded. If \code{TRUE}, then the \code{xvar} and
\code{yvar} transformations from the original model fit are excluded from
the \code{optimize}. From example, if original model is fit with \code{xvar
  = log} and \code{yvar = NULL}, then the \code{optimize} is translated into
\code{optimize = list(x = c('NULL', 'sqrt'),
 y = c('log',  'sqrt'))}.}

\item{add_fit_criteria}{An optional (default \code{NULL}) indicator to add
fit criteria to the model fit. options are \code{loo} and \code{waic}.
Please see \code{\link[brms:add_criterion]{brms::add_criterion()}} for details.}

\item{add_fit_bayes_R}{An optional (default \code{NULL}) to add Bayesian R
square.}

\item{...}{Other arguments passed to \code{\link{brms}}.}
}
\value{
An updated object of class \code{brmsfit, bsiatr}, that contains the
posterior draws and other useful information about the model.
}
\description{
Optimize \pkg{bsitar} model
}
\examples{
\dontrun{
data(heights)
data_males <- heights \%>\% filter(sex == 'Male')
fit_males <- bsitar(x=age, y=height, id=id, data=heights, df=4)
fit_males2 <- optimize_bsitar(fit_males)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
