% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check_bgm.R
\name{pp_check_bgm.bgmfit}
\alias{pp_check_bgm.bgmfit}
\alias{pp_check_bgm}
\title{Posterior predictive distribution check for \code{bgmfit} object}
\usage{
\method{pp_check_bgm}{bgmfit}(model, resp = NULL, envir = parent.frame(), ...)

pp_check_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bgmfit} function for details).}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}
function. Please see \code{brms::pp_check.brmsfit} for details on
various options available.}
}
\value{
A ggplot object that can be further customized using the
ggplot2 package.
}
\description{
Perform posterior predictive checks with the help
of the \pkg{bayesplot} package.
}
\details{
The \code{pp_check_bgm} function is a wrapper around
the \link[brms:loo.brmsfit]{brms::loo} and works exactly the same way as
\link[brms:pp_check.brmsfit]{brms::pp_check.brmsfit}.
}
\examples{
#
# Fit Bayesian SITAR model 
# berkeley_fit <- bgm(x = age, y = height, id = id, data = data, df = 4,
#                     chains = 2, iter = 1000, thin = 10)
#
# To avoid running the model which takes some time, the fitted model has 
# already been saved as berkeley_fit.rda object. The model is fitted using 2 
# chain  with 1000  iteration per chain (to save time) and setting thin as 1 
# (to save memory also).
# 
model <- berkeley_fit
#
pp_check_bgm(model)

}
