% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_.bsitar.R
\name{predict_.bsitar}
\alias{predict_.bsitar}
\alias{predict_}
\title{Predicted values (draws) from the posterior predictive distribution}
\usage{
\method{predict_}{bsitar}(model, resp = NULL, deriv = 0, envir = parent.frame(), ...)

predict_(model, ...)
}
\arguments{
\item{model}{An object of class \code{bsitar}.
function.}

\item{resp}{Response variable (default \code{NULL}) specified as a string
character required during the post-processing of multivariate and
univariate-by-subgroup model (see \code{bsitar} function for details).}

\item{deriv}{An integer value to specify whether to estimate distance curve
(i.e., model estimated curve(s)) or velocity curve (first derivative of the
model estimated curve(s)). A 0 value (default) is for distance curve and
1 for the velocity curve.}

\item{envir}{The calling environment. Deafault set to \code{parent.frame()}.}

\item{...}{Additional arguments passed to the \link[brms:predict.brmsfit]{brms::predict.brmsfit}
function. Please see \code{brms::predict.brmsfit} for details on
various options available.}
}
\value{
An array of predicted response values. See \link[brms:predict.brmsfit]{brms::predict.brmsfit}
for details.
}
\description{
This method is an alias of \code{\link[brms]{posterior_predict.brmsfit}}
with additional arguments for obtaining summaries of the computed draws.
}
\details{
The \code{predict_} function is a wrapper around
the \link[brms:predict.brmsfit]{brms::predict.brmsfit}. The \link[brms:predict.brmsfit]{brms::predict.brmsfit}
function from the \code{brms} package can used to plot the predict (distance)
curve for an \emph{bsitar} model when outcome (e.g., height) is untransformed.
However, when the outcome is log or square root transformed, the
\link[brms:predict.brmsfit]{brms::predict.brmsfit} will return the predict curve on the log or
square root scale whereas the \link{predict_} will
return the predict curve on the original scale. Furthermore, the
predict_ also displays the velocity curve on the original scale
after making required back-transformation. Apart from these differences,
both these functions (\link[brms:predict.brmsfit]{brms::predict.brmsfit} and
\link{predict_}) work in the same manner. In other words,
user can specify all the arguments which are available in the
\link[brms:predict.brmsfit]{brms::predict.brmsfit}. Because of this, the name is kept same except
for adding an underscore at the end of the name (\emph{predict} to
\emph{predict_}).
}
\examples{
\dontrun{
# The examples below show the use of *predict_* to estimate  
# population average and individual-specific distance and velocity 
# curves for the the predict model
# Population average distance curve
predict_(model, deriv = 0, re_formula = NA)

# Individual-specific distance curves
predict_(model, deriv = 0, re_formula = NULL)

# Population average velocity curve
predict_(model, deriv = 1, re_formula = NA)

# Individual-specific velocity curves
predict_(model, deriv = 1, re_formula = NULL)
 
}
}
