% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_function.R
\name{prepare_function}
\alias{prepare_function}
\title{Prepare Stan function for Bayesian SITAR growth curve model}
\usage{
prepare_function(x, y, id, knots, nknots, data, internal_function_args)
}
\arguments{
\item{x}{Predictor variable in the data. See \code{\link[=bgm]{bgm()}} for details.}

\item{y}{Response variable in the data. See \code{\link[=bgm]{bgm()}} for details.}

\item{id}{A vector specifying a unique group identifier for each individual.
See \code{\link[=bgm]{bgm()}} for details.}

\item{knots}{A vector of knots used for constructing the spline design
matrix. See \code{\link[=bgm]{bgm()}} for details.}

\item{nknots}{An integer specifying the number of knots.}

\item{data}{Data frame containing variables \code{x}, \code{y} and \code{id}.}

\item{internal_function_args}{Internal arguments passed from the
\code{\link[=bgm]{bgm()}} to the \code{prepare_formula}).}
}
\value{
An character string which later evaluated to a custom function
and inserted into the Stan's functions block.
}
\description{
The \code{prepare_function}) constructs custom Stan function  which is passed
on to the \code{\link[=bgm]{bgm()}} function. For univariate-by- subgroup model
(\code{univariate_by}) and multivariate (\code{multivariate}) models (see
\code{\link[=bgm]{bgm()}}), the \code{x}, \code{y}, \code{id}, \code{knots},
\code{nknots}, are automatically matched with the sub-models.
}
