% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors_initials.R
\name{set_priors_initials}
\alias{set_priors_initials}
\title{Set priors and initials for fitting Bayesian SITAR growth curve model}
\usage{
set_priors_initials(
  a_prior_beta,
  b_prior_beta,
  c_prior_beta,
  d_prior_beta,
  s_prior_beta,
  a_cov_prior_beta,
  b_cov_prior_beta,
  c_cov_prior_beta,
  d_cov_prior_beta,
  s_cov_prior_beta,
  a_prior_sd,
  b_prior_sd,
  c_prior_sd,
  d_prior_sd,
  a_cov_prior_sd,
  b_cov_prior_sd,
  c_cov_prior_sd,
  d_cov_prior_sd,
  gr_prior_cor,
  rsd_prior_sigma,
  dpar_prior_sigma,
  dpar_cov_prior_sigma,
  autocor_prior_acor,
  mvr_prior_rescor,
  prior_data,
  prior_data_internal,
  prior_args_internal,
  init_arguments,
  init_data,
  init_data_internal,
  init_args_internal
)
}
\arguments{
\item{a_prior_beta}{Set priors on the the fixed effect \code{a} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{a_prior_beta} for details.}

\item{b_prior_beta}{Set priors on the the fixed effect \code{b} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{b_prior_beta} for details.}

\item{c_prior_beta}{Set priors on the the fixed effect \code{c} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{c_prior_beta} for details.}

\item{d_prior_beta}{Set priors on the the fixed effect \code{d} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{d_prior_beta} for details.}

\item{s_prior_beta}{Set priors on the the fixed effect \code{s} parameter
(i.e., spline coeficients). See \code{\link[=bsitar]{bsitar()}} function,
\code{s_prior_beta} for details.}

\item{a_cov_prior_beta}{Set priors on the the covariate(s) for the fixed
effect \code{a} parameter. See \code{\link[=bsitar]{bsitar()}} function,
\code{a_cov_prior_beta} for details.}

\item{b_cov_prior_beta}{Set priors on the the covariate(s) for the fixed
effect \code{b} parameter. See \code{\link[=bsitar]{bsitar()}} function,
\code{b_cov_prior_beta} for details.}

\item{c_cov_prior_beta}{Set priors on the the covariate(s) for the fixed
effect \code{c} parameter. See \code{\link[=bsitar]{bsitar()}} function,
\code{c_cov_prior_beta} for details.}

\item{d_cov_prior_beta}{Set priors on the the covariate(s) for the fixed
effect \code{d} parameter. See \code{\link[=bsitar]{bsitar()}} function,
\code{d_cov_prior_beta} for details.}

\item{s_cov_prior_beta}{Set priors on the the covariate(s) for the fixed
effect \code{s} parameter (i.e., spline coeficients).
See \code{\link[=bsitar]{bsitar()}} function, \code{s_cov_prior_beta} for details.}

\item{a_prior_sd}{Set priors on the the group effect (i.e., random) effect
for \code{a} parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{a_prior_sd}
for details.}

\item{b_prior_sd}{Set priors on the the group effect (i.e., random) effect
for \code{b} parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{b_prior_sd}
for details.}

\item{c_prior_sd}{Set priors on the the group effect (i.e., random) effect
for \code{c} parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{c_prior_sd}
for details.}

\item{d_prior_sd}{Set priors on the the group effect (i.e., random) effect
for \code{d} parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{d_prior_sd}
for details.}

\item{a_cov_prior_sd}{Set priors on the the covariate(s) for the
group effect (i.e., random) effect for \code{a} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{a_cov_prior_sd} for details.}

\item{b_cov_prior_sd}{Set priors on the the covariate(s) for the
group effect (i.e., random) effect for \code{b} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{b_cov_prior_sd} for details.}

\item{c_cov_prior_sd}{Set priors on the the covariate(s) for the
group effect (i.e., random) effect for \code{c} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{c_cov_prior_sd} for details.}

\item{d_cov_prior_sd}{Set priors on the the covariate(s) for the
group effect (i.e., random) effect for \code{d} parameter.
See \code{\link[=bsitar]{bsitar()}} function, \code{a_cov_prior_sd} for details.}

\item{gr_prior_cor}{Set priors on the the group level correlation
parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{d_prior_beta} for details.}

\item{rsd_prior_sigma}{Set priors on the the residual standared deviation
\code{sigma} parameter. See \code{\link[=bsitar]{bsitar()}} function, \code{a_prior_beta}
for details.}

\item{dpar_prior_sigma}{Set priors on the distributional \code{sigma}
parameter (which is same as residual standared deviation for Gaussian
distribution). See \code{\link[=bsitar]{bsitar()}} function, \code{a_prior_beta}
for details.}

\item{dpar_cov_prior_sigma}{Set priors on the covariates for the
distributional \code{sigma} parameter (which is same as residual standard
deviation for Gaussian distribution). See \code{\link[=bsitar]{bsitar()}} function,
\code{a_prior_beta} for details.}

\item{autocor_prior_acor}{Set priors on the the autocorrelation parameters
\code{ar}, \code{ma} and \code{arma}. See \code{\link[=bsitar]{bsitar()}} function,
\code{a_prior_beta} for details.}

\item{mvr_prior_rescor}{Set priors on the the residual correlation
parameter for multivariate model. See \code{\link[=bsitar]{bsitar()}} function,
\code{d_prior_beta} for details.}

\item{prior_data}{An optional argument as named list to pass value for prior.
See \code{\link[=bsitar]{bsitar()}} function, \code{prior_data} for details.}

\item{prior_data_internal}{An internal data frame (named list) to pass on
the relevant information on priors from the \code{\link[=bsitar]{bsitar()}} function to
the \code{set_priors_initials}.}

\item{prior_args_internal}{An internal argument list that is passed from the
\code{\link[=bsitar]{bsitar()}} function to the \code{set_priors_initials} and is used
for setting the priors.}

\item{init_arguments}{A list containing all the init arguments specified in
the \code{\link[=bsitar]{bsitar()}} function and now passed on to the
\code{set_priors_initials}.}

\item{init_data}{An optional data argument (named list) used to pass initial
values. See \code{\link[=bsitar]{bsitar()}} function, \code{prior_data} for details.}

\item{init_data_internal}{An internal data frame (named list) to pass on
the relevant information on initials from the \code{\link[=bsitar]{bsitar()}} function to
the \code{set_priors_initials}.}

\item{init_args_internal}{An internal argument list that is passed from the
\code{\link[=bsitar]{bsitar()}} function to the \code{set_priors_initials} and is used
for setting the initials.}
}
\value{
An object of class \code{brmsprior} (See \code{brmsprior}).
In addition to the priors, the returned object contains a list of initial
values.
}
\description{
The \code{set_priors_initials}) sets priors and initials values which are
passed from the \link{bsitar} function to \code{set_priors_initials}.
For univariate-by- subgroup model (specified by using the
\code{univariate_by}) and multivariate model (specified by using the
\code{multivariate}), each argument is automatically matched with
the sub-model(s).
}
