% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bgm.R
\name{update_bgm.bgmfit}
\alias{update_bgm.bgmfit}
\alias{update_bgm}
\title{Update a model of class \code{bgmfit}.}
\usage{
\method{update_bgm}{bgmfit}(model, newdata = NULL, recompile = NULL, ...)

update_bgm(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{newdata}{An optional \code{data.frame} to be used when updating the
model. If \code{NULL} (default), the same data used in the original model
fit is evaluated. Note that data-dependent default priors will not be
updated automatically.}

\item{recompile}{A logical to indicate whether the Stan model should be
recompiled. When \code{NULL} (the default), \code{bsitar::update_bgm()}
tries to figure out internally, if recompilation is necessary. Setting it
to \code{FALSE} will cause all Stan code changing arguments to be ignored.}

\item{...}{Other arguments passed to \code{\link{brms}}.}
}
\value{
An updated model of class \code{brmsfit}, that contains the
posterior draws and other useful information about the model.
}
\description{
Update a model of class \code{bgmfit}.
}
\details{
This is an adapted version of the \code{update} function from the
\code{brms} package.
}
\examples{
\dontrun{
# Fit Bayesian SITAR model 
# berkeley_fit <- bgm(x = age, y = height, id = id, data = data, df = 4,
#                     chains = 2, iter = 1000, thin = 10)

# To avoid running the model which takes some time, the fitted model has 
# already been saved as berkeley_fit.rda object. The model is fitted using 2 
# chain  with 1000  iteration per chain (to save time) and setting thin as 1 
# (to save memory also).
# 
fit_males <- bgm(x=age, y=height, id=id, data=heights, df=4)
model2 <- update_bgm(model, df = 5)
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
