% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model.R
\name{update_model.bgmfit}
\alias{update_model.bgmfit}
\alias{update_model}
\title{Update model}
\usage{
\method{update_model}{bgmfit}(model, newdata = NULL, recompile = NULL, ...)

update_model(model, ...)
}
\arguments{
\item{model}{An object of class \code{bgmfit}.}

\item{newdata}{An optional \code{data.frame} to be used when updating the
model. If \code{NULL} (default), the same data used in the original model
fit is evaluated. Note that data-dependent default priors will not be
updated automatically.}

\item{recompile}{A logical to indicate whether the Stan model should be
recompiled. When \code{NULL} (the default), \code{bsitar::update_model()}
tries to figure out internally, if recompilation is necessary. Setting it
to \code{FALSE} will cause all Stan code changing arguments to be ignored.}

\item{...}{Other arguments passed to \code{\link{brms}}.}
}
\value{
An updated model of class \code{brmsfit}, that contains the
posterior draws and other useful information about the model.
}
\description{
Update model
}
\details{
This is an adapted version of the \strong{update} function from the
\pkg{brms} package.
}
\examples{

\donttest{
# Fit Bayesian SITAR model 
# To avoid running the model which takes some time, model fit to the
# \code{berkeley_mdata} has already been saved as berkeley_mfit.rda object.
# Please see \code{bgm} examples.

model <- berkeley_mfit

# To save time, below example is fit with sample_prior = 'only'

model2 <- update_model(model, df = 5, iter = 100, sample_prior = 'only')
}

}
\author{
Satpal Sandhu  \email{satpal.sandhu@bristol.ac.uk}
}
